### 2025-01-05 #7628 @danielskatz
👋 @sergei-mironov - thanks for your submission.  As it's not clear to me that this is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS, the editors will now discuss this. You should hear back in a week or two.

If this does pass this review, I will then ask you to reduce the length of the paper to closer to 1000 words, the JOSS maximum. There also appear to a be couple of reference/DOI issues that we will also discuss.
