+++ paper/md/paper.md
@@ -21,7 +21,12 @@ bibliography: paper.bib
 Litrepl is a lightweight text processing tool designed to recognize and evaluate
 code sections within Markdown or Latex documents. This functionality is useful
 for both batch document section evaluation and interactive coding within a text
-editor, provided a straightforward integration is established.
+editor, provided a straightforward integration is established. Inspired by
+Project Jupyter, it aims to facilitate the creation of research documents. In
+the light of recent developments in software deployment, however, we have
+shifted our focus from ad-hoc reproducibility to enhancing transparency in
+communication with programming language interpreters, by either eliminating or
+clearly exposing mutable states within the communication process.
 
 # Statement of need
 
