+++ paper/tex/paper.bib
@@ -30,12 +30,23 @@
               language-dependence, feature complexity, documentation, code extraction}
 }
 @article{Perez2007IPython,
-  title={IPython: A System for Interactive Scientific Computing},
-  author={Fernando P{\'e}rez and Brian E. Granger},
-  journal={Computing in Science \& Engineering},
-  year={2007},
-  volume={9},
-  url={https://api.semanticscholar.org/CorpusID:16637923}
+  author = {Perez, Fernando and Granger, Brian E.},
+  title = {IPython: A System for Interactive Scientific Computing},
+  year = {2007},
+  issue_date = {May 2007},
+  publisher = {IEEE Educational Activities Department},
+  address = {USA},
+  volume = {9},
+  number = {3},
+  issn = {1521-9615},
+  url = {https://doi.org/10.1109/MCSE.2007.53},
+  doi = {10.1109/MCSE.2007.53},
+  abstract = {Python offers basic facilities for interactive work and a comprehensive library on top of which more sophisticated systems can be built. The IPython project provides an enhanced interactive environment that includes, among other features, support for data visualization and facilities for distributed and parallel computation.},
+  journal = {Computing in Science and Engg.},
+  month = may,
+  pages = {21–29},
+  numpages = {9},
+  keywords = {scientific programming, scientific computing, computer languages, Python}
 }
 @article{Granger2021litcomp,
   author={Granger, Brian E. and Pérez, Fernando},
