+++ paper/tex/paper.tex
@@ -62,7 +62,7 @@ clearly exposing mutable states within the communication process.
 
 \begin{figure*}[!hbt]
   \centering
-  \includesvg[inkscapelatex=false,width=0.8\textwidth]{pic}
+  \includegraphics{pic}
   \caption{Litrepl resource allocation diagram. Hash \textbf{A} is computed
   based on the Litrepl working directory and the interpreter class. Hash
   \textbf{B} is computed based on the contents of the code section.}
