+++ Makefile
@@ -93,7 +93,7 @@ pq: ./paper/tex/paper_quick.pdf
 .PHONY: p
 paper: ./paper/tex/paper.pdf
 p: ./paper/tex/paper.pdf
-./paper/tex/paper.pdf: ./paper/tex/paper_checked.tex ./paper/tex/pic.svg ./paper/tex/paper.bib
+./paper/tex/paper.pdf: ./paper/tex/paper_checked.tex ./paper/tex/pic.png ./paper/tex/paper.bib
 	cd paper/tex && \
 	latexmk -shell-escape -pdf -latex=pdflatex paper_checked.tex && \
 	cp paper_checked.pdf paper.pdf
@@ -107,7 +107,7 @@ p: ./paper/tex/paper.pdf
 .PHONY: paper.tar.gz
 paper.tar.gz: $(PAPER_TAR_GZ)
 $(PAPER_TAR_GZ): ./paper/tex/paper.pdf
-	tar -czvf $@ -C paper/tex paper.tex pic.svg paper.bib paper.bbl
+	tar -czvf $@ -C paper/tex paper.tex pic.png paper.bib paper.bbl
 
 
 .PHONY: all
