+++ paper/tex/paper.tex
@@ -62,7 +62,7 @@ clearly exposing mutable states within the communication process.
 
 \begin{figure*}[!hbt]
   \centering
-  \includegraphics{pic}
+  \includegraphics{pic.pdf}
   \caption{Litrepl resource allocation diagram. Hash \textbf{A} is computed
   based on the Litrepl working directory and the interpreter class. Hash
   \textbf{B} is computed based on the contents of the code section.}
