+++ README.md
@@ -18,6 +18,12 @@ plugin to demonstrate editor integration.
 
 </div>
 
+
+_Notes:_
+* _[literate programming](https://en.wikipedia.org/wiki/Literate_programming)_
+* _[read-eval-print-loop coding](https://en.wikipedia.org/wiki/Read%E2%80%93eval%E2%80%93print_loop)_
+
+
 <table border="0">
   <tr>
     <td>
@@ -27,17 +33,11 @@ plugin to demonstrate editor integration.
     </td>
     <td>
       Preprint: <i>(2025, Sergei Mironov)</i>
-      <span style="text-decoration: underline;">
-        Litrepl: Literate Paper Processor Promoting Transparency More Than Reproducibility
-      </span>
+      <ins>Litrepl: Literate Paper Processor Promoting Transparency More Than Reproducibility</ins>
       <a href="https://arxiv.org/abs/2501.10738">arXiv:2501.10738</a>
     </td>
   </tr>
 </table>
-
-_Notes:_
-* _[literate programming](https://en.wikipedia.org/wiki/Literate_programming)_
-* _[read-eval-print-loop coding](https://en.wikipedia.org/wiki/Read%E2%80%93eval%E2%80%93print_loop)_
 
 🔥 Features
 -----------
