### 2025-01-24 bce831f
Reducing the md paper length
