### 2025-01-24 fe01dc0
Reducing the md paper length
