+++ paper/md/paper.md
@@ -48,47 +48,28 @@ programming, allowing users to modify the interpreter state directly. This
 approach is believed to enhance human thought processes by keeping users
 actively involved [@Granger2021litcomp].
 
-A significant milestone in this field was the IPython interpreter
-[@Perez2007IPython], which later evolved into the Jupyter Project. Its creators
-introduced a new document format called the Jupyter Notebook [@Kluyver2016jupnb],
-characterized by a series of logical sections of various types, including text
-and code, which could directly interact with programming language interpreters.
-This interactive communication, akin to REPL style programming, allows the
-creation of well-structured documents suitable for presentations and sharing.
-The concept underpinning these developments is termed *Literate Computing*
-[@Perez2015blog], which includes goals of spanning a wide audience range,
-boosting reproducibility, and fostering collaboration. To achieve these
-objectives, several technical decisions were made, notably the introduction of
-bidirectional communication between the computational core, known as the Jupyter
-Kernel, and the Notebook web-based renderer, along with another layer of
-client-server communication between the web-server and the user’s web browser.
+A pivotal development in this field was the IPython interpreter
+[@Perez2007IPython], which led to the Jupyter Project and its Jupyter Notebook
+format [@Kluyver2016jupnb]. This format consists of logical sections like text
+and code that can interact with language interpreters, enabling REPL-like
+programming for well-structured, shareable documents. This is part of *Literate
+Computing* [@Perez2015blog], which aims to reach broad audiences, enhance
+reproducibility, and promote collaboration. Key technical aspects include
+bidirectional communication between the Jupyter Kernel and Notebook renderer,
+alongside client-server interactions between the web server and user browser.
 
-While we recognize the importance of all goals within the Literate Computing
-framework, we think that the goal of reproducibility is more important than
-others. Addressing it alone would suffice to enhance communication among
-time-separated and space-separated researchers and significantly expand the
-audience. However, as it became clear [@Dolstra2010], this challenge extends
-beyond the scope of a single human-computer interaction system, and even beyond
-the typical boundaries of software distribution management for a particular
-programming language. A comprehensive solution to the software deployment
-problem operates at the entire operating system level.
-
-Following [@Vallet2022], we propose shifting human-computer interaction towards
-simplicity and transparency by introducing *Litrepl*. This tool offers
-REPL-style editing, integrates with existing editors and text formats, and
-minimizes hidden state variables, all while maintaining a compact codebase.
-Litrepl uses bidirectional text streams for inter-process communication with
-language interpreters and supports existing text formats like Markdown and LaTeX
-through simplified parsers. It currently supports Python, Shell, and a custom
-large language model interpreter, aiming to utilize POSIX [@POSIX2024] system
-features extensively.
+We believe that reproducibility is crucial in the Literate Computing framework,
+enhancing communication among dispersed researchers. However, as noted by
+[@Dolstra2010], we argue that this challenge exceeds the capacity of a single
+system, needing operating system-level solutions. Inspired by [@Vallet2022], we
+introduce *Litrepl*, which enables REPL-style editing by integrating with existing
+editors and formats, and reduces hidden state variables in a compact codebase.
+Litrepl uses bidirectional text streams for inter-process communication and
+supports Markdown and LaTeX formats with simplified parsers. It supports Python,
+Shell, and a custom large language model communication interpreter while
+leveraging POSIX [@POSIX2024] system features.
 
 # How it works
-
-Litrepl is implemented as a command-line text utility. Its primary function is
-to take a text document as input, process it according to specified command-line
-arguments and environment settings, and then output the resulting document
-through its standard output.
 
[…]
