### 2025-01-24 cd8033e
Reducing the md paper length
