### 2025-01-30 056d849
Print section begin and end regexprs
