### 2025-02-06 #7628 @danielskatz
@sergei-mironov - any news on the shortened paper? I will continue looking for an editor as well, and we have some new ones coming on board soon...
