### 2025-02-06 #7628 @fabian-s
@danielskatz I'd rather not, sorry.

I did not vote on scope of this submission, but I would definitely not have supported taking this on and would rather spend my time on submissions I'm more interested in. It also doesn't seem fair to the submitter to have an editor that doesn't see the point of the paper from the outset, since we aim for a collaborative, constructive review process at JOSS.

To me, this looks like an extremely feature-limited early prototype version of things like Quarto/JupyterNotebook/Rmarkdown/Sweave/.... and I don't see why anybody would choose to use this instead of these much more mature frameworks for literate programming.  (no offense, @sergei-mironov)
