### 2025-02-06 3f173e4
Add print-auxdir command
