### 2025-02-10 427ce2e
Move ipython to sh/
