+++ python/litrepl/grammar.py
@@ -36,7 +36,10 @@ def markdown_com_sec(tag:str, st:SectionType=SectionType.Code)->SectionGrammar:
   return SectionGrammar(f'com{tag}', f'<!--[ ]*{tag}[ ]*-->', f'<!--[ ]*no{tag}[ ]*-->', st)
 
 def latex_sec(tag:str, st:SectionType=SectionType.Code)->SectionGrammar:
-  return SectionGrammar(tag, '\\begin{%s}'%(tag,), '\\end{%s}'%(tag,), st)
+  return SectionGrammar(tag, r'\\begin{%s}'%(tag,), r'\\end{%s}'%(tag,), st)
+
+def latex_com_sec(tag:str, st:SectionType=SectionType.Code)->SectionGrammar:
+  return SectionGrammar(f'com{tag}', f'\%[ ]*{tag}[ ]*\%', f'\%[ ]*no{tag}[ ]*\%', st)
 
 def secbody(sg:SectionGrammar)->str:
   return dedent(f'''
@@ -61,6 +64,7 @@ def mk_markdown_grammar(tags:list[str]|None=None)->str:
   tags=tags or []
   code_sections=[
     markdown_sec('python', SectionType.Code),
+    markdown_com_sec('python', SectionType.Code),
     *[markdown_sec(tag, SectionType.Code) for tag in tags],
     *[markdown_com_sec(tag, SectionType.Code) for tag in tags]
   ]
@@ -68,6 +72,21 @@ def mk_markdown_grammar(tags:list[str]|None=None)->str:
     markdown_sec('result', SectionType.Result),
     markdown_com_sec('result', SectionType.Result),
     markdown_com_sec('ignore', SectionType.Ignore),
+  ]
+  return mkgrammar([*code_sections,*other_sections])
+
+def mk_latex_grammar(tags:list[str]|None=None)->str:
+  tags=tags or []
+  code_sections=[
+    latex_sec('python', SectionType.Code),
+    latex_com_sec('python', SectionType.Code),
+    *[latex_sec(tag, SectionType.Code) for tag in tags],
+    *[latex_com_sec(tag, SectionType.Code) for tag in tags]
+  ]
+  other_sections=[
+    latex_sec('result', SectionType.Result),
+    latex_com_sec('result', SectionType.Result),
+    latex_com_sec('ignore', SectionType.Ignore),
   ]
   return mkgrammar([*code_sections,*other_sections])
 
