+++ python/litrepl/base.py
@@ -34,6 +34,9 @@ from .eval import (process, pstderr, rresultLoad, rresultSave, processAdapt,
                    with_fd, read_nonblock, eval_code, eval_code_)
 from .utils import(unindent, indent, escape, fillspaces, fmterror,
                    cursor_within, nlines, wraplong, remove_silent, hashdigest)
+from .grammar import (SectionType, st2rule, SectionGrammar, markdown_sec,
+                      markdown_com_sec, latex_sec, secbody, mkgrammar,
+                      mk_markdown_grammar, mk_latex_grammar)
 
 from .interpreters.python import PythonInterpreter
 from .interpreters.ipython import IPythonInterpreter
@@ -265,11 +268,11 @@ class SymbolsMarkdown(Symbols):
     self.ignorebegin=r"<!--[ ]*l?ignore[ ]*-->"
     self.ignoreend=r"<!--[ ]*l?noignore[ ]*-->"
     self.secbegin_dict={
-      'vim':(codebegin_re+'|'+comcodebegin_re).replace('?','\?'),
+      'vim':(codebegin_re+'|'+comcodebegin_re).replace('?',r'\?'),
       'lark':(codebegin_re+'|'+comcodebegin_re)
     }
     self.secend_dict={
-      'vim':(self.resultend+'|'+self.comresultend).replace('?','\?'),
+      'vim':(self.resultend+'|'+self.comresultend).replace('?',r'\?'),
       'lark':(self.resultend+'|'+self.comresultend),
     }
 
@@ -314,89 +317,112 @@ CBR="}"
 BCBR="\\}"
 BOBR="\\{"
 
+
+# def grammar_(a:LitreplArgs, filetype:str)->Tuple[LarkGrammar,Symbols]:
+#   # For the `?!` syntax, see
+#   # https://stackoverflow.com/questions/56098140/how-to-exclude-certain-possibilities-from-a-regular-expression
+#   if filetype in ["md","markdown"]:
+#     symbols_md=SymbolsMarkdown(a)
+#     sl=symbols_md
+#     toplevel_markers_markdown='|'.join([
+#       sl.codebegin,sl.resultbegin,
+#       sl.comresultbegin, sl.ignorebegin, sl.comcodebegin
+#     ])
+#     return (dedent(fr"""
+#       start: (topleveltext)? (snippet (topleveltext)?)*
+#       snippet : codesec -> e_icodesection
+#               | resultsec -> e_ocodesection
+#               | ignoresec -> e_comsection
+#       ignoresec.2 : ignorebegin ignoretext ignoreend
+#       codesec.1 : codebegin codetext codeend
+#                      | comcodebegin codesectext comcodeend
+#       resultsec.1 : resultbegin resulttext resultend
+#                      | comresultbegin vertext comresultend
+#       codebegin : /{symbols_md.codebegin}/
+#       codeend : /{symbols_md.codeend}/
+#       comcodebegin : /{symbols_md.comcodebegin}/
+#       comcodeend : /{symbols_md.comcodeend}/
+#       resultbegin : /{symbols_md.resultbegin}/
+#       resultend : /{symbols_md.resultend}/
+#       comresultbegin : /{symbols_md.comresultbegin}/
+#       comresultend : /{symbols_md.comresultend}/
+#       inlinebeginmarker : "`"
+#       inlinendmarker : "`"
+#       ignorebegin : /{symbols_md.ignorebegin}/
+#       ignoreend : /{symbols_md.ignoreend}/
+#       topleveltext : /(.(?!{toplevel_markers_markdown}))*./s
[…]
