### 2025-02-11 19b2714
Integrating modular grammar
