### 2025-02-12 #7628 @sergei-mironov
> [@sergei-mironov](https://github.com/sergei-mironov) - any news on the shortened paper? I will continue looking for an editor as well, and we have some new ones coming on board soon...

@danielskatz  I have shortened the paper text. For the [last version](https://github.com/sergei-mironov/litrepl/blob/paper/paper/md/paper.md), if I remove the yaml header and a few technical comments, my word counter would show 990. Please let me know if you want it to be even shorter.
