### 2025-02-12 #7628 @danielskatz
👋 @sergei-mironov - thanks. Can you now add something brief that compares this to other tools, as this is one of the JOSS review criteria for the paper (State of the field: Do the authors describe how this software compares to other commonly-used packages?) In particular, it would be good to compare with things like Jupyter, Quarto, Rmarkdown, ...  I know this will lengthen the paper a bit, but a small bit more is ok, as this is what reviewers will need to have.
