### 2025-02-13 0a2d921
Fix spurious newline on empty inputs
