### 2025-02-13 5f22f1d
Fix spurious newline on empty inputs
