+++ README.md
@@ -3,23 +3,14 @@
 ⌨️ LitRepl ⌨️
 </h1>
 
-[Changelog](./CHANGELOG.md) | [Installation](https://sergei-mironov.github.io/litrepl/installation) | [Usage](https://sergei-mironov.github.io/litrepl/usage/general-concepts/) | [Gallery](#-gallery)
-
-[![](./img/coverage.svg)](./docs/coverage.md)
+[Changelog](./CHANGELOG.md) | [Installation](#-installation) | [Usage](#basic-execution) | [Gallery](#-gallery)
 </div>
 
-<!--
-``` python
-!echo -n Litrepl is a
-!cat docs/static/description.md | sed '/^\#/d;/^$/d'
-```
--->
-<!--result-->
-Litrepl is acommand-line processor for *Markdown* or *LaTeX* documents with
+**Litrepl** is a command-line processor for *Markdown* or *LaTeX* documents with
 **literate programming** code sections. Instructed by its arguments, it
 evaluates and updates sections via background interpreters. Interpreters can
-stay active for a **read-eval-paste-loop** style.
-<!--noresult-->
+stay active for a **read-eval-paste-loop** style. The repository includes a Vim
+plugin to demonstrate editor integration.
 
 <div align="center">
 
@@ -48,13 +39,13 @@ _Notes:_
   </tr>
 </table>
 
-Features
---------
+🔥 Features
+-----------
 
 * **Document formats** <br/>
-  Markdown _(Example [[MD]](./docs/examples/example.md))_ **|**
+  Markdown _(Example [[MD]](./doc/example.md))_ **|**
   [LaTeX](https://www.latex-project.org/)
-  _(Examples [[TEX]](./docs/examples/example.tex)[[PDF]](./docs/examples/example.pdf))_
+  _(Examples [[TEX]](./doc/example.tex)[[PDF]](./doc/example.pdf))_
 * **Interpreters** <br/>
   [Sh](https://en.wikipedia.org/wiki/Bourne_shell) **|**
   [Bash](https://www.gnu.org/software/bash/) **|**
@@ -64,8 +55,8 @@ Features
 * **Editor integration** <br/>
   [Vim](https://www.vim.org/scripts/script.php?script_id=6117) _(plugin source included)_
 
-Requirements
-------------
+✅ Requirements
+---------------
 
 * **POSIX-compatible OS**, typically a Linux. The tool relies on POSIX
   operations, notably pipes, and depends on certain Shell commands.
@@ -73,21 +64,949 @@ Requirements
 * **[Socat](http://www.dest-unreach.org/socat/)** (Optional) Needed for
   `litrepl repl` and Vim's `LTerm` commands to work.
 
-Documentation
--------------
+⚖️ Comparison
+------------
 
-* [Installation](https://sergei-mironov.github.io/litrepl/installation/)
-* [Basic usage](https://sergei-mironov.github.io/litrepl/usage/general-concepts/)
-* [Formatting documents](https://sergei-mironov.github.io/litrepl/usage/formatting/)
+| Name        | Document formats            | Interpreters    | Backend API  | Frontend API |
+|-------------|-----------------------------|-----------------|--------------|--------------|
+| Jupyter     | Jupyter[1]                  | Many(>50)[2]    | Server       | Web, IDE     |
+| Quarto      | Markdown+Quarto[3], Jupyter | Few(3)[4]       | Server       | Web, IDE     |
+| Codebraid   | Markdown+Codebraid[5]       | Many(>7,>50)[6] | Server       | IDE          |
+| R markdown  | RMarkdown[7]                | Many(>8)[8]     | Server       | IDE          |
[…]
