+++ README.md
@@ -67,31 +67,36 @@ _Notes:_
 ⚖️ Comparison
 ------------
 
-| Name        | Document formats            | Interpreters    | Backend API  | Frontend API |
-|-------------|-----------------------------|-----------------|--------------|--------------|
-| Jupyter     | Jupyter[1]                  | Many(>50)[2]    | Server       | Web, IDE     |
-| Quarto      | Markdown+Quarto[3], Jupyter | Few(3)[4]       | Server       | Web, IDE     |
-| Codebraid   | Markdown+Codebraid[5]       | Many(>7,>50)[6] | Server       | IDE          |
-| R markdown  | RMarkdown[7]                | Many(>8)[8]     | Server       | IDE          |
-| Litrepl     | Markdown, LaTeX             | Few(3)          | Pipes        | Command line |
+| Name        | Document formats   | Data formats | Interpreters    | Backend                      | Frontend             |
+|-------------|--------------------|--------------|-----------------|------------------------------|----------------------|
+| Jupyter     | Jupyter[j1]        | Rich         | Many(>50)[j2]   | Client-server via ZeroMQ[j3] | Web, Editor          |
+| Quarto      | Markdown+[q1]      | Rich         | Few             | Modified Jupyter[q2]         | Web, Editor          |
+| Codebraid   | Markdown+[c1]      | Rich         | Many[c2]        | Same as Jupyter[c3]          | Editor               |
+| R markdown  | Markdown+[r1]      | Rich         | Few[r2]         | Linked libraries[r3]         | Editor               |
+| Litrepl     | Markdown, LaTeX    | Text-only    | Few             | POSIX Pipes                  | Command line, Editor |
 
 
 * [Jupyter](https://jupyter.org/)
-  - [1] [Jupyter Notebook Format](https://nbformat.readthedocs.io/en/latest/)
-  - [2] [Available kernels](https://github.com/jupyter/jupyter/wiki/Jupyter-kernels)
-  - [Kernel](https://docs.jupyter.org/en/latest/projects/kernels.html)
+  - [j1] [Jupyter Notebook Format](https://nbformat.readthedocs.io/en/latest/)
+  - [j2] [Available kernels](https://github.com/jupyter/jupyter/wiki/Jupyter-kernels)
+  - [j3] Details on Jupyter backend communication [link](https://docs.jupyter.org/en/latest/projects/kernels.html)
 * [Quarto](https://quarto.org/docs/tools/text-editors.html)
-  - [3] [Quarto Markdown extensions](https://quarto.org/docs/computations/execution-options.html)
-  - [4] [Jupyter kernel support](https://quarto.org/docs/advanced/jupyter/kernel-execution.html)
+  - [q1] [Quarto Markdown extensions](https://quarto.org/docs/computations/execution-options.html)
+  - [q2] Quarto uses the slightly modified Jupyter kernel protocol. See
+    [Jupyter kernel support](https://quarto.org/docs/advanced/jupyter/kernel-execution.html)
+    [Qurto echo kernel](https://github.com/quarto-dev/quarto_echo_kernel/blob/b77fde70c25a869175cc225a3676eee2df5a1733/README.rst)
 * [Codebraid](https://codebraid.org/)
-  - [5] [Code chunks formatting](https://codebraid.org/code_chunks/)
-  - [6] [Codebraid languages](https://github.com/gpoore/codebraid/blob/3f85800bd58c2a1587778a1fe0f24c46dc1c3a69/README.md?plain=1#L13-L16)
+  - [c1] Codebraid Markdown extensions are described as
[…]
