### 2025-02-25 #7628 @danielskatz
@sergei-mironov - note that you can do this too...  editorialbot commands need to be the first entry in a new comment.
