### 2025-02-25 #7628 @danielskatz
@xuanxu & @arfon - Can we handle footnotes/references in a table? Do you a suggestion for what's best in this case? And should we add something in our example paper to explain this case?

> [@danielskatz](https://github.com/danielskatz) The original text compared my solution with Jupyter as it is probably the most popular and mature approach. I also have added the brief comparison with other projects including the ones you requested, in form of a table. Here is the [direct link](https://github.com/sergei-mironov/litrepl/blob/paper/paper/md/paper.md?plain=1#L77) pointing to the text. I might need some advises on how to format the references within this table. The ideas is to have table-local footnotes, if the publishing system allows it. Is it possible to express it in the JOSS Markdown dialect? Please let me know if you want me to make further updates.
