### 2025-02-25 #7628 @xuanxu
> Can we handle footnotes/references in a table? Do you a suggestion for what's best in this case? And should we add something in our example paper to explain this case?

Probably @tarleb can answer here
