### 2025-02-26 #7628 @tarleb
Footnotes in tables become normal, document-wide footnotes.

Maybe [definition lists](https://pandoc.org/MANUAL#definition-lists) might be an option here?
