### 2025-03-25 #7628 @matthewfeickert
@sergei-mironov Thanks for your patience. Though before I spend reviewer time on this I'm going to request that you reread the [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html) documentation and also suggest that you read the [Review criteria docs](https://joss.readthedocs.io/en/latest/review_criteria.html) to better understand how your submission would be evaluated.

At the moment, the documentation is rather weak as there is only a README and a few other Markdown documents. In general we're looking for much more robust documentation.

While I appreciate that Litrepl is meant to be a CLI tool/text editor plugin only, and not be used as a Python library, we still expect full API documentation with examples. From simply running `litrepl --help` it can be seen that is not provided.

The tests consist of a single Shell script, but it isn't clear what coverage this provides, and as you're using external Python dependencies outside of the CPython standard library there should be automated tests for [all Python versions Litrepl supports](https://github.com/sergei-mironov/litrepl/blob/5f22f1d3cd899c5330d1c37affd4deb682b7d228/setup.py#L63). Use with modern Python versions raises `SyntaxWarning`s that should have been caught by automated testing.

I am also not very convinced by the statement of need and examples as to how Litrepl is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). The [table in the paper](https://github.com/sergei-mironov/litrepl/blob/14f302022959855564be8085eea3c00729e21661/paper/md/paper.md?plain=1#L77-L95) (which needs to have a Table number as a reference and a caption) seems to also not be making a strong case for Litrepl, but that's a decision that should be left to reviewers. Though please make the arguments very clear for them.

At the moment the submission does not feel like it is in a complete state to have reviewers actively engage with it. It should be clear to the reviewers exactly what they are evaluating through the documentation, sufficient test set, and clear statement of need.

If you would like to continue with the review at this stage, please address all the above issues within two weeks (by 2025-04-09).

If that timeline isn't sufficient or viable for your current availability, we can negotiate on that, but if you feel like you would like more time to fully address everything, my suggestion there, given the scope of work that would need to be done, is that you instead withdraw your submission for now and resubmit in the future whenever you have had the time to carefully go through and prepare the project. That's totally fine, not uncommon, and future (re)submissions would be welcome.
