### 2025-03-28 #7628 @sergei-mironov
Hi @matthewfeickert , thank you for paying attention to this submission. I appreciate your points; let me first comment on the scope issue, and then I will answer the technical ones.

> I am also not very convinced by the statement of need and examples as to how Litrepl is research software.

I admit that the Litrepl scope might be considered a corner case by some people.  I tend to call it research software because publishing research results is a valuable part of the process, and I believe that Litrepl facilitates the delivery of these results from the computation core to output documents in the most direct and observable manner. I would like to argue politely that according to the modern definition of 'reproducibility', neither Litrepl nor other tools (not even Jupyter, despite the claims of its authors) make papers reproducible by themselves; they need to be part of a reproducible research environment to achieve this. Litrepl, however, seems to me the lightest option by far for those who are ready to cope with its limitations. I would like to try passing the review process to publish these arguments.

> The documentation is rather weak as there is only a README and a few other Markdown documents

Sure, I admit that the Readme file is the document that currently contains the most information. I am ready to provide documents in other formats too. I think that for a CLI application, a manpage might be a good addition. Please let me know which format you find appropriate.

> we still expect full API documentation with examples

I did not originally plan to provide Python documentation because I did not consider this project to be a Python library (so, strictly speaking, there is no public API). However, the latest changes introduced a few distinct modules which might be interesting to a wider audience, namely the interpreter wrapper classes and the text document grammar combinators. I am ready to provide the documentation; just let me know which format you find best. Would you consider very detailed code comments as good documentation, or should it be a Read the Docs page?

> The tests consist of a single Shell script, but it isn't clear what coverage this provides

The single shell script you mentioned contains approximately 30 integration test scenarios, and some of them are reused with different interpreters. I test many aspects of parsing and evaluation, including integration with the Vim editor I currently maintain. Since the project relies on POSIX APIs, I think that wrapping POSIX calls in standard Python tests would be unnecessary.

I agree that measuring the coverage would be a good thing to do, so I am ready to integrate the codecov tools. Thank you for mentioning the SyntaxWarning issue; I will fix it as soon as possible, along with adding an option to test more Python versions.

> If you would like to continue with the review at this stage, please address all the above issues within two weeks (by 2025-04-09).

I'll be busy for the next two weeks, but after that, before the middle of April, I hope to have plenty of free time to work on this project, including improving the tests and adding the documentation. Would it be okay if we extended the deadline by a week? I will do my best to resolve the issues before April 16, 2025. What do you think?
