### 2025-04-11 752121a
Fix backward-compatibility issues
