### 2025-04-11 f85a28e
Fix SyntaxWarning in Python3.12
