### 2025-04-11 fed2a74
Add test shell environment featuring multiple python versions
