+++ sh/runtests.sh
@@ -1406,7 +1406,8 @@ runlitrepl() {
   test -n "$LITREPL_TEST_AI_INTERPRETER"
   test -n "$LITREPL_TEST_SH_INTERPRETER"
   test -n "$LITREPL_BIN"
-  $LITREPL_BIN/litrepl --debug="$LITREPL_DEBUG" \
+  test -n "$LITREPL_TEST_PYTHON"
+  $LITREPL_TEST_PYTHON $LITREPL_BIN/litrepl --debug="$LITREPL_DEBUG" \
     --python-interpreter="$LITREPL_TEST_PYTHON_INTERPRETER" \
     --ai-interpreter="$LITREPL_TEST_AI_INTERPRETER" \
     --sh-interpreter="$LITREPL_TEST_SH_INTERPRETER" \
@@ -1431,9 +1432,11 @@ Usage: runtest.sh [-d] [-i I(,I)*] [-t T(,T)*]
 Arguments:
   -d                        Be very verbose
   -i I, --interpreters=I    Run tests requiring interpreters matching the grep expression I
-                            Run -i '?' to list all available interpreters.
+                            Run -i '?' to list all available interpreters
   -t T, --tests=T           Run tests whose names match the grep expression T
-                            Run -t '?' to list all available tests.
+                            Run -t '?' to list all available tests
+  -p P, --python=P          Use this Python interpreter to run Litrepl
+                            Run -p '?' to list available python interpreters
 
 Examples:
   runtests.sh -t '?' -i '?'
@@ -1453,14 +1456,19 @@ while test -n "$1" ; do
   case "$1" in
     --interpreters=*) INTERPS=$(echo "$1" | sed 's/.*=//g') ;;
     -i|--interpreters) INTERPS="$2"; shift ;;
-    --tests=) TESTS=$(echo "$1" | sed 's/.*=//g') ;;
+    --tests=*) TESTS=$(echo "$1" | sed 's/.*=//g') ;;
     -t|--tests) TESTS="$2"; shift ;;
+    --python=*) LITREPL_TEST_PYTHON=$(echo "$1" | sed 's/.*=//g') ;;
+    -p|--python) LITREPL_TEST_PYTHON="$2"; shift ;;
     -d|-V|--verbose) set -x; LITREPL_DEBUG=1 ;;
     -h|--help) usage ; exit 1 ;;
   esac
   shift
 done
 
+if test "$LITREPL_TEST_PYTHON" = "?" ; then
+  which --all python
+fi
 if test "$TESTS" = "?" ; then
   tests | awk '{print $1}' | sort -u
 fi
@@ -1469,11 +1477,14 @@ if test "$INTERPS" = "?" ; then
   tests | awk '{print $3}' | grep -v -w '-' | sort -u
   tests | awk '{print $4}' | grep -v -w '-' | sort -u
 fi
-if test "$INTERPS" = "?" -o "$TESTS" = "?" ; then
+if test "$INTERPS" = "?" -o "$TESTS" = "?" -o "$LITREPL_TEST_PYTHON" = "?" ; then
   exit 1
 fi
 if test -z "$LITREPL_BIN"; then
   LITREPL_BIN=$LITREPL_ROOT/python/bin
+fi
+if test -z "$LITREPL_TEST_PYTHON" ; then
+  LITREPL_TEST_PYTHON=python
 fi
 
 trap "echo FAIL\(\$?\)" EXIT
