### 2025-04-13 357d2d2
Update env: Add coverage-badge dependency
