+++ sh/runtests.sh
@@ -1437,7 +1437,8 @@ Arguments:
                             Run -t '?' to list all available tests
   -p P, --python=P          Use this Python interpreter to run Litrepl
                             Run -p '?' to list available python interpreters
-  -c FILE, --coverage=FILE  Collect coverage results into the FILE
+  -c FILE, --coverage=FILE  Collect coverage results into the FILE. Defaults to \`.coverage\`.
+  -c -,    --coverage=-     Disable coverage collecting
 
 Examples:
   runtests.sh -t '?' -i '?'
@@ -1454,6 +1455,9 @@ unset AICLI_HISTORY
 INTERPS='.*'
 TESTS='.*'
 LITREPL_DEBUG=0
+if test -z "$LITREPL_COVERAGE" ; then
+  LITREPL_COVERAGE=$(pwd)/.coverage
+fi
 while test -n "$1" ; do
   case "$1" in
     --interpreters=*) INTERPS=$(echo "$1" | sed 's/.*=//g') ;;
@@ -1470,11 +1474,7 @@ while test -n "$1" ; do
   shift
 done
 
-case "$LITREPL_COVERAGE" in
-  ""|/*) ;;
-  *) LITREPL_COVERAGE="$(pwd)/$LITREPL_COVERAGE" ;;
-esac
-
+# Queries handling
 if test "$LITREPL_TEST_PYTHON" = "?" ; then
   which --all python
 fi
@@ -1489,6 +1489,18 @@ fi
 if test "$INTERPS" = "?" -o "$TESTS" = "?" -o "$LITREPL_TEST_PYTHON" = "?" ; then
   exit 1
 fi
+
+# Coverage setup
+case "$LITREPL_COVERAGE" in
+  ""|/*) ;;
+  -) unset LITREPL_COVERAGE ;;
+  *) LITREPL_COVERAGE="$(pwd)/$LITREPL_COVERAGE" ;;
+esac
+if test -n "$LITREPL_COVERAGE" ; then
+  coverage erase
+fi
+
+# "Binary" setup
 if test -z "$LITREPL_BIN"; then
   LITREPL_BIN=$LITREPL_ROOT/python/bin
 fi
@@ -1516,4 +1528,9 @@ test "$NRUN" = "0" && die "No tests were run" || true
 )
 trap "" EXIT
 echo OK
+
+if test -n "$LITREPL_COVERAGE" ; then
+  coverage combine ${LITREPL_COVERAGE}.*
+  coverage-badge -o img/coverage.svg
+fi
 
