### 2025-04-14 c42f900
Update tests: handle the shell-script wrapper case
