### 2025-04-14 38bdc48
Add .ondirrc
