### 2025-04-17 2231771
Update env: Collect coverage only when making full testing
