### 2025-04-17 6ef5efc
Update status command: print auxdir, fix unwanted exception
