+++ README.md
@@ -5,7 +5,7 @@
 
 [Changelog](./CHANGELOG.md) | [Installation](#-installation) | [Usage](#basic-execution) | [Gallery](#-gallery)
 
-![](./img/coverage.svg)
+[![](./img/coverage.svg)](#-coverage-report)
 </div>
 
 **Litrepl** is a command-line processor for *Markdown* or *LaTeX* documents with
@@ -99,6 +99,7 @@ _Notes:_
     * [Tools for Screencast Recording](#tools-for-screencast-recording)
     * [Common Development Scenarios](#common-development-scenarios)
 * [🎥 Gallery](#-gallery)
+* [🖼️ Coverage report](#-coverage-report)
 * [💡 Technical Insights](#-technical-insights)
 * [🚫 Known Limitations](#-known-limitations)
 * [Related Tools and Projects](#related-tools-and-projects)
@@ -754,12 +755,12 @@ usage: litrepl [-h] [-v] [--filetype STR] [--python-markers STR[,STR]]
                [--foreground] [--map-cursor LINE:COL:FILE]
                [--result-textwidth NUM]
               
-{start,stop,restart,status,parse,parse-print,eval-sections,eval-code,repl,interrupt,print-regexp,print-grammar}
+{start,stop,restart,status,parse,parse-print,eval-sections,eval-code,repl,interrupt,print-regexp,print-grammar,print-auxdir}
                ...
 
 positional arguments:
  
-{start,stop,restart,status,parse,parse-print,eval-sections,eval-code,repl,interrupt,print-regexp,print-grammar}
+{start,stop,restart,status,parse,parse-print,eval-sections,eval-code,repl,interrupt,print-regexp,print-grammar,print-auxdir}
                               Commands to execute
     start                     Start the background interpreter.
     stop                      Stop the background interpreters.
@@ -780,6 +781,7 @@ positional arguments:
                               the given file type.
     print-grammar             Print the resulting grammar for the given
                               filetype.
+    print-auxdir              Print the auxdir for the given interpreter type.
 
 options:
   -h, --help                  show this help message and exit
@@ -964,9 +966,15 @@ Usage: runtest.sh [-d] [-i I(,I)*] [-t T(,T)*]
 Arguments:
   -d                        Be very verbose
   -i I, --interpreters=I    Run tests requiring interpreters matching the grep expression I
-                            Run -i '?' to list all available interpreters.
+                            Run -i '?' to list all available interpreters
   -t T, --tests=T           Run tests whose names match the grep expression T
-                            Run -t '?' to list all available tests.
+                            Run -t '?' to list all available tests
+  -p P, --python=P          Use this Python interpreter to run Litrepl
+                            Run -p '?' to list available python interpreters
+  -c FILE, --coverage=FILE  Collect coverage results into the FILE. Defaults to
+                            `.coverage` if no tests or interpreters are
+                            selected, otherwize disabled.
+  -c -,    --coverage=-     Disable coverage.
 
 Examples:
   runtests.sh -t '?' -i '?'
@@ -1002,6 +1010,32 @@ preview LaTeX documents instantly.
 <video controls src="https://user-images.githubusercontent.com/4477729/187065835-3302e93e-6fec-48a0-841d-97986636a347.mp4" muted="true"></video>
 
 </details>
+
+📊 Coverage report
+------------------
+
+``` sh
+coverage report --format=markdown -m
+```
+
+<!--result-->
+| Name                                        |    Stmts |     Miss |   Cover |   Missing |
+|-------------------------------------------- | -------: | -------: | ------: | --------: |
[…]
