### 2025-04-17 4e3a72a
Update Makefile: generate coverage report
