+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 <!--
 vim: tw=80
---!>
+-->
 Changelog
 =========
 
