### 2025-04-18 #7628 @sergei-mironov
Let me post an update on this issue. At the time of the writing I have improved the Litrepl tool in the following ways:

* [x] Integrated the code coverage metrics, it turned out to be approximately 88%. The detailed report is available in the corresponding Readme file section.
* [x] Improved the test system so it can test Litrepl using differetn Python interpreters. Fixed a number of compatibility issues.
* [x] Fixed the SyntaxWarning when running tests using newer Pythons

There are some requests that remains undone. I plan to address them within a couple of weeks.

* [ ] Write API documentation
* [ ] Write a manpage (Is it really a requirement?)
