### 2025-04-18 c03c63d
Add initial manpage building rule
