### 2025-04-18 dc12194
Add initial manpage building rule
