### 2025-04-18 35bed96
Add initial manpage building rule
