### 2025-04-21 be41984
Add initial manpage building rule
