### 2025-04-22 #7628 @matthewfeickert
This is a form notifcation to all JOSS work that I'm involved with editing at the moment:

Sorry for being slow to respond to comments here and follow up. I've been sick and then had work travel and so am behind on all JOSS work. Please ping me if you do not have a more full response to ongoing work within 24 hours.

If you have questions or concerns, please feel free to ask for additional information.
