### 2025-04-25 a88cc90
Update vim: drop g:litrepl_bin and g:litrepl_exe, read LITREPL, set VIM_LITREPL_FILE
