+++ python/litrepl/types.py
@@ -5,32 +5,40 @@ from dataclasses import dataclass
 from enum import Enum
 from lark import ParseTree as LarkTree
 
+# String containing a file name.
 FileName=str
+
+# Litrepl command-line arguments. See `litrepl.main` for the declaration.
 LitreplArgs=Any
+
+# Lark parser grammar.
 LarkGrammar=str
 
-# Process exit code or None (running)
+# Process exit code or None which means that the process is still running.
 ECode=Optional[int]
 ECODE_RUNNING=None
 ECODE_OK=0
 ECODE_UNDEFINED=255
 
 class SType(Enum):
-  """ Code section types. They have short string representation and allows user
-  to use short code section names and specify interpreter paths and parameters
-  via settings."""
+  """ Code section types, also known as `interpreter class`. In combination with
+  interpreter path, encode the `Interpreter` descendant used for handling the
+  session. See also `litrepl.base.start`. """
   SPython = 0
   SAI = 1
   SShell = 2
 
 @dataclass(frozen=True)
 class RunResult:
-  """ Result of launchng the readout job """
+  """ Result of launchng the readout job. Every section execution results in
+  creating a readout job process which reads the output and writes it into a
+  temporary file `fname`."""
   fname:FileName     # File where the output data is piped into
 
 @dataclass
 class ReadResult:
-  """ Result of reading from the readout job """
+  """ Result of reading from the readout job, as specified by a `RunResult`
+  structure. """
   text:str           # Current contents of the readout file.
   timeout:bool       # Did the current read attmept timeout? If so, Litrepl
                      # would return control to the user with a
@@ -73,17 +81,19 @@ class PrepInfo:
 
 @dataclass
 class SecRec:
-  """ Request for section evaluation """
+  """ Request for section evaluation in an input document. """
   nsecs:Set[NSec]                   # Sections to evaluate
   preproc:PrepInfo                  # Results of preprocessing
 
   @staticmethod
   def empty():
+    """ Create the default empty `SecRec`."""
     return SecRec(set(),PrepInfo.empty())
 
 @dataclass
 class FileNames:
-  """ Interpreter state """
+  """ Background interpreter session state. Encodes the path and the contents of
+  the auxiliary directory `wd`. """
   wd:str                            # Working directory
   inp:str                           # Input pipe
   outp:str                          # Output pipe
@@ -98,8 +108,9 @@ SECVAR_RE = re_compile("(\^+ *R[0-9]+ *\^+)|(v+ *R[0-9]+ *v+)|(\>+ *R[0-9]+ *\<+
 
 @dataclass
 class EvalState:
-  """ Interpreter state, tracking evaluation of document sections """
-  sr:SecRec                         # The original request
+  """ Litrepl interpreter state, trackes the processing of a document passed as
+  Litrepl intput. """
+  sr:SecRec                         # The request for sections evaluation
   sres:Dict[int,str]                # Section results: sec.num -> result
   ledder:Dict[int,int]              # Facility to restore the cursor: line -> offset
   ecodes:Dict[int,ECode]            # Exit codes: sec.num -> exitcode
@@ -111,35 +122,39 @@ class EvalState:
 
 
 class Interpreter:
-  """ Interpreter abstraction. """
+  """ Third-party interpreter base. Encodes the low-level POSIX API which
+  Litrepl requires to start and manage the (possibly background) session. See
+  the `litrepl.interpreters` submodule for implementations. """
   def __init__(self, fns:FileNames)->None:
     """ Create the interpreter object, associated with certain files, as
     specified by `fns`."""
     self.fns=fns
-  def run_child(self,interpreter:str)->int:
+  def run_child(self, interpreter:str)->int:
     """ Launch the interpreter process and wait for it to complete, then return
     the system's exit code. This method should create only one child process.
[…]
