### 2025-04-28 6342ce5
Fix man section 5->1; Fixed remaining help TODOs
