### 2025-04-28 bce8131
Update inline help formatting
