+++ python/litrepl/main.py
@@ -44,52 +44,63 @@ def make_parser():
     help='Specify the type of input formatting (markdown|[la]tex|auto).')
   ap.add_argument('--python-markers',metavar='STR[,STR]',
     default=environ.get('LITREPL_PYTHON_MARKERS','python'),
-    help=dedent('''Specify section markers recognized as `python` sections.
+    help=dedent('''
+    Specify section markers recognized as `python` sections.
     Defaults to the value of LITREPL_PYTHON_MARERS if set, otherwize
     "python".'''))
   ap.add_argument('--ai-markers',metavar='STR[,STR]',
     default=environ.get('LITREPL_AI_MARKERS','codeai,ai'),
-    help=dedent('''Specify section markers recognized as `ai` sections.
+    help=dedent('''
+    Specify section markers recognized as `ai` sections.
     Defaults to the value of LITREPL_AI_MARERS if set, otherwize
     "codeai,ai".'''))
   ap.add_argument('--sh-markers',metavar='STR[,STR]',
     default=environ.get('LITREPL_SH_MARERS','sh,bash'),
-    help=dedent('''Specify section markers recognized as `shell` sections.
+    help=dedent('''
+    Specify section markers recognized as `shell` sections.
     Defaults to the value of LITREPL_SH_MARERS if set, otherwize "shell". '''))
   ap.add_argument('--python-interpreter',metavar='EXE',
     default=_ensure_nonepty(environ.get('LITREPL_PYTHON_INTERPRETER','auto')),
-    help=dedent('''Python interpreter command line, or `auto`. Defaults to the
+    help=dedent('''
+    Python interpreter command line, or `auto`. Defaults to the
     LITREPL_PYTHON_INTERPRETER environment variable if set, otherwise "auto".
     Litrepl determines "python" or "ipython" type according to the value.'''))
   ap.add_argument('--ai-interpreter',metavar='EXE',
     default=_ensure_nonepty(environ.get('LITREPL_AI_INTERPRETER','auto')),
-    help=dedent('''`aicli` interpreter command line or `auto`. Defaults to the
+    help=dedent('''
+    `aicli` interpreter command line or `auto`. Defaults to the
     LITREPL_AI_INTERPRETER environment variable if set, otherwise "auto".'''))
   ap.add_argument('--sh-interpreter',metavar='EXE',
     default=_ensure_nonepty(environ.get('LITREPL_SH_INTERPRETER','auto')),
-    help=dedent('''Shell interpreter command line or `auto`. Defaults to the
+    help=dedent('''
+    Shell interpreter command line or `auto`. Defaults to the
     LITREPL_SH_INTERPRETER environment variable if set, otherwise "auto".'''))
   ap.add_argument('--python-auxdir',type=str,metavar='DIR',
     default=_ensure_nonepty(environ.get('LITREPL_PYTHON_AUXDIR')),
-    help=dedent('''This directory stores Python interpreter pipes. It defaults
+    help=dedent('''
+    This directory stores Python interpreter pipes. It defaults
     to LITREPL_PYTHON_AUXDIR if set; otherwise, it's created in the system's
     temporary directory, named after the current working directory.'''))
   ap.add_argument('--ai-auxdir',type=str,metavar='DIR',
     default=_ensure_nonepty(environ.get('LITREPL_AI_AUXDIR')),
-    help=dedent('''This directory stores AI interpreter pipes. It defaults to
+    help=dedent('''
+    This directory stores AI interpreter pipes. It defaults to
     LITREPL_AI_AUXDIR if set; otherwise, it's created in the system's temporary
     directory, named after the current working directory.'''))
[…]
