### 2025-04-28 d7e03d8
Update man: add project description
