### 2025-04-28 8003974
Add mkdocs support
