+++ docs/usage/general-concepts.md
@@ -1,0 +1,314 @@
+## General concepts
+
+The Litrepl tool identifies code and result sections within a text document. It
+processes the code by sending it to the appropriate interpreters and populates
+the result sections with their responses. The interpreters remain active in the
+background, ready to handle new inputs.
+
+Litrepl supports subsets of **Markdown** and **LaTeX** formatting in order to
+recognize the sections. Some aspects of the recognized grammars, such as section
+labels, could be configured.
+
+#### Basic Execution
+
+Litrepl searches for verbatim code sections followed by zero or more result
+sections. In Markdown documents, the Python code is any triple-quoted section
+with a pre-configured label such as `python`. The result is any triple-quoted
+`result` section.  In LaTeX documents, sections are marked with
+`\begin{...}\end{...}` environments correspondingly.
+
+The primary command for evaluating formatted documents is `litrepl
+eval-sections`. Consider a [hello.md](../static/hello.md) document:
+
+<!--
+``` python
+print("~~~~ markdown")
+!cat docs/static/hello.md
+print("~~~~")
+```
+-->
+
+<!-- result -->
+~~~~ markdown
+``` python
+print('Hello Markdown!')
+```
+``` result
+Hello Markdown!
+```
+~~~~
+<!-- noresult -->
+
+You pass it to Litrepl using:
+
+~~~~ shell
+$ cat hello.md | litrepl eval-sections > result.md
+~~~~
+
+The `result.md` will have all sections filled in correctly.
+<!--
+``` python
+print("~~~~ markdown")
+!cat docs/static/hello.md | litrepl --foreground eval-sections
+print("~~~~")
+```
+-->
+
+<!-- result -->
+~~~~ markdown
+``` python
+print('Hello Markdown!')
+```
+``` result
+Hello Markdown!
+```
+~~~~
+<!-- noresult -->
+
+* For additional details on Markdown formatting, refer to [Formatting Markdown
+  documents](./formatting.md#markdown)
+
+Similarly, in a LaTeX document, we format code and result sections as follows:
+
+~~~~ tex
+\begin{python}
+print('Hello LaTeX!')
+\end{python}
+
+\begin{result}
+Hello LaTeX!
+\end{result}
+~~~~
+
+- LaTeX documents require a preamble introducing python/result environments to
+  the TeX processor. For more information, see [Formatting LaTeX
+  documents](./formatting.md#latex).
+
+By default, Litrepl tried to guess the format of the input document. Use the
+`--filetype=(latex|markdown)` option to set the format explicitly:
+
+``` sh
+$ cat doc.md | litrepl --filetype=markdown eval-sections
+$ cat doc.tex | litrepl --filetype=latex eval-sections
+```
+
+* The main Vim command for code section evaluation is `:LEval`. By default, it
+  executes the section at the cursor. To execute all sections in a document, use
+  `:LEval all`.
+
+#### Selecting Sections for Execution
+
+By default, `litrepl eval-sections` evaluates all sections in a document. To
+evaluate only specific sections, the range argument should be specified. The
+overall syntax is `litrepl eval-sections [RANGE]`, where `RANGE` can be:
+
+* `N`: Represents a specific code section to evaluate, with the following
+  possible formats:
+  - A number starting from `0`.
+  - `$` symbol, indicating the last section.
+  - `L:C`, referring to the line and column position. Litrepl calculates the
+    section number based on this position.
+* `N..N`: Represents a range of sections, determined using the rules mentioned
+  above.
+
+Some examples:
+
+``` sh
+$ litrepl eval-sections '0'       # First section in a document
+$ litrepl eval-sections '3..$'    # Sections from fourth section (zero based) to the last one
[…]
