+++ docs/static/hello.md
@@ -1,0 +1,6 @@
+``` python
+print('Hello Markdown!')
+```
+``` result
+Hello Markdown!
+```

[…]
