+++ docs/index.md
@@ -1,17 +1,7 @@
-# Welcome to MkDocs
+# Litrepl
 
-For full documentation visit [mkdocs.org](https://www.mkdocs.org).
-
-## Commands
-
-* `mkdocs new [dir-name]` - Create a new project.
-* `mkdocs serve` - Start the live-reloading docs server.
-* `mkdocs build` - Build the documentation site.
-* `mkdocs -h` - Print help message and exit.
-
-## Project layout
-
-    mkdocs.yml    # The configuration file.
-    docs/
-        index.md  # The documentation homepage.
-        ...       # Other markdown pages, images and other files.
+**Litrepl** is a command-line processor for *Markdown* or *LaTeX* documents with
+**literate programming** code sections. Instructed by its arguments, it
+evaluates and updates sections via background interpreters. Interpreters can
+stay active for a **read-eval-paste-loop** style. The repository includes a Vim
+plugin to demonstrate editor integration.
