+++ docs/about.md
@@ -1,70 +1,0 @@
-# About Litrepl
-
-## Caede ire
-
-Lorem markdownum semihomines antistita vix nec Dryope vultum prohibetis, vide
-celer velis. Puro constitit at magna obvertit dies veritus utque, nunc
-[carere](http://data-pedes.com/)! Peneide nova nati phaedimus coeptus eque
-perfudit sola flexit. Medio gemitus traderet durata. Protinus primoque summis:
-*has*, comaeque: scilicet lacusque accipit mariti.
-
-## Procorum tempusque
-
-Nisi a morsu ducemque vastator captatus, mota arsit discutiunt audaci veste
-creatis [exitus fossas](http://www.orbemsi.net/illic), est! Ut Iphide, esse
-maris cedere memorantur [inponere
-iamque](http://colorque-cognoscere.org/virgosimul). Nullo eloquio vestes, voce
-*gelidas feri*. Et parabat vobis perosus feret aut, cum ad aestibus. Illic
-laborum Libycas lugentis dissiluit cupidine eripiunt minor Cithaeron retices me,
-loqui cerva sibi repertis qua hanc qualem plura.
-
-    dvd_page_services = rw.throughput_toggle_it(htmlOleFirmware) - 902709 +
-            time(2, nybble);
-    if (bankruptcyPath + alpha - clock) {
-        ciscClobWeb.im_ad -= ecc;
-    }
-    zif_access = windowsVleArt;
-    if (scanProgramming(telecommunications, archive_website(dslamCpaHorse, 4,
-            bus_hard), remote_zero_pcmcia - boot) != quad +
-            joystick_opengl_web) {
-        pointSessionDesktop -= jreFrameworkBar(kernelRwDtd);
-        payloadDrive.firmware_spyware_screenshot = 19 + cpaTraceroute;
-    }
-
-## Admissumque qualia vacuaque alumno pectus
-
-*Ara* sucis elusam **tangeret** noviens. Dicenda dicentem amor addit decipit
-antiquus.
-
-    if (thick) {
-        e -= streaming(mirrored, drop_beta_trackback);
-    } else {
-        supercomputerSnowSpool = bugRtf.backlink_online_inkjet(characterSnippet,
-                jquery_telnet_interface * -2);
-        dvd(5);
-        down.raster_sql_dvd -= 45;
-    }
-    cache(5 + mca_gpu, symbolic_ram_windows + token_white_pipeline.formula(
-            pharmingVfatAntivirus, point_domain_activex));
-    if (91) {
-        play_data += cacheFlatExabyte;
-    }
-    macintosh = fileLogic(module_gnu + favoritesUgcTemplate, uddiNanometerUdp,
-            system_activex_ram);
-
-## Cohaesit expulit est
-
-Sic ducis nitar. Me exitus: cupidine, pertulerint nunc adfusaeque et neque
-*iungere monilia*. Nox stirpe, Aconteus iactura hic mori cum vade, ira? Auditque
-nisi aequos it rumpere Cyane.
-
-> Moratur amplectens et nulla Achilles ab laetabitur discedens habenas
-> sequeretur me [teneri quaere roganda](http://per.com/sumereut.html), pater
-> deducit; pes nec. Mihi includit ambitiosa Nemeaeo parenti potius pollutosque
-> vidit auditque? Torva fuerit et neque, animasse eburnea; est terram rursus ceu
-> et frui properas Orionis redolentia metum.
-
-Breviore hunc terga quam egens subiere: intravit caelatus vicit pectore vocat
-et. Esse triumphos nec sublimis casus dominus neque ceciderat: sui viridem ipse
[…]
