+++ docs/static/hello.md
@@ -2,5 +2,4 @@
 print('Hello Markdown!')
 ```
 ``` result
-Hello Markdown!
 ```
