+++ docs/usage/general-concepts.md
@@ -34,7 +34,6 @@ print("~~~~")
 print('Hello Markdown!')
 ```
 ``` result
-Hello Markdown!
 ```
 ~~~~
 <!-- noresult -->
