+++ mkdocs.yml
@@ -1,0 +1,11 @@
+site_name: Litrepl Docs
+nav:
+  - Home: index.md
+  - Installation: installation.md
+  - Usage:
+    - General concepts: usage/general-concepts.md
+    - Formatting: usage/formatting.md
+    - Application Scenarios: usage/application-scenarios.md
+  - Reference: reference.md
+  - Development: development.md
+theme: readthedocs
