+++ README.md
@@ -137,80 +137,3 @@ coverage report --format=markdown -m
 |                                   **TOTAL** | **1453** |  **241** | **83%** |           |
 <!--noresult-->
 
-💡 Technical Insights
----------------------
-
-The following events should normally happen after users type the `:LitEval1`
-command:
-
-1. On the first run, LitREPL starts the Python interpreter in the background.
-   Its standard input and output are redirected into UNIX pipes in the current
-   directory.
-2. LitREPL runs the whole document through the express Markdown/Latex parser
-   determining the start/stop positions of code and result sections. The cursor
-   position is also available and the code from the right code section can
-   reach the interpreter.
-3. The process which reads the interpreter's response is forked out of the main
-   LitREPL process. The output goes to the temporary file.
-4. If the interpreter reports the completion quickly, the output is pasted to
-   the resulting document immediately. Otherwise, the temporary results are
-   pasted.
-5. Re-evaluating sections with temporary results causes LitREPL to update
-   these results.
-
-🚫 Known Limitations
---------------------
-
-* Formatting: Nested code sections are not supported.
-* ~~Formatting: Special symbols in the Python output could invalidate the
-  document~~.
-* Interpreter: Extra newline is required after Python function definitions.
-* Interpreter: Stdout and stderr are joined together.
-* ~~Interpreter: Evaluation of a code section locks the editor~~.
-* Interpreter: Tweaking `os.ps1`/`os.ps2` prompts of the Python interpreter
-  could break the session.
-* ~~Interpreter: No asynchronous code execution.~~
-* ~~Interpreter: Background Python interpreter couldn't be interrupted~~
-
-Related Tools and Projects
---------------------------
-
-Edititng:
-
-* https://github.com/lervag/vimtex (LaTeX editing, LaTeX preview)
-* https://github.com/shime/vim-livedown (Markdown preview)
-* https://github.com/preservim/vim-markdown (Markdown editing)
-
-Code execution:
-
-* Vim-medieval https://github.com/gpanders/vim-medieval
-  - Evaluates Markdown code sections
-* Pyluatex https://www.ctan.org/pkg/pyluatex
-* Magma-nvim https://github.com/dccsillag/magma-nvim
-* Codi https://github.com/metakirby5/codi.vim
-* Pythontex https://github.com/gpoore/pythontex
-  - Evaluates Latex code sections
-* Codebraid https://github.com/gpoore/codebraid
-* Vim-ipython-cell https://github.com/hanschen/vim-ipython-cell
-* Vim-ipython https://github.com/ivanov/vim-ipython
-* Jupytext https://github.com/goerz/jupytext.vim
-  - Alternative? https://github.com/mwouts/jupytext
-* Ipython-vimception https://github.com/ivanov/ipython-vimception
-
-Useful Vim plugins:
-
-* https://github.com/sergei-grechanik/vim-terminal-images (Graphics in vim terminals)
-
-Useful tools:
-
-* https://pandoc.org/
-
-Considerations for Third-Party Tools
-------------------------------------
-
-* Vim-plug https://github.com/junegunn/vim-plug/issues/1010#issuecomment-1221614232
-* Pandoc https://github.com/jgm/pandoc/issues/8598
-* Jupytext https://github.com/mwouts/jupytext/issues/220#issuecomment-1418209581
-* Vim-LSC https://github.com/natebosch/vim-lsc/issues/469
-* [Bad PDF fonts in Firefox](https://github.com/mozilla/pdf.js/issues/17401)
-
