+++ docs/related.md
@@ -1,0 +1,32 @@
+# Related
+## Tools and Projects
+
+Edititng:
+
+* <https://github.com/lervag/vimtex> (LaTeX editing, LaTeX preview)
+* <https://github.com/shime/vim-livedown> (Markdown preview)
+* <https://github.com/preservim/vim-markdown> (Markdown editing)
+
+Code execution:
+
+* Vim-medieval <https://github.com/gpanders/vim-medieval>
+  - Evaluates Markdown code sections
+* Pyluatex <https://www.ctan.org/pkg/pyluatex>
+* Magma-nvim <https://github.com/dccsillag/magma-nvim>
+* Codi <https://github.com/metakirby5/codi.vim>
+* Pythontex <https://github.com/gpoore/pythontex>
+  - Evaluates Latex code sections
+* Codebraid <https://github.com/gpoore/codebraid>
+* Vim-ipython-cell <https://github.com/hanschen/vim-ipython-cell>
+* Vim-ipython <https://github.com/ivanov/vim-ipython>
+* Jupytext <https://github.com/goerz/jupytext.vim>
+  - Alternative? <https://github.com/mwouts/jupytext>
+* Ipython-vimception <https://github.com/ivanov/ipython-vimception>
+
+## Considerations for Third-Party Tools
+
+* Vim-plug <https://github.com/junegunn/vim-plug/issues/1010#issuecomment-1221614232>
+* Pandoc <https://github.com/jgm/pandoc/issues/8598>
+* Jupytext <https://github.com/mwouts/jupytext/issues/220#issuecomment-1418209581>
+* Vim-LSC <https://github.com/natebosch/vim-lsc/issues/469>
+
