+++ README.md
@@ -5,7 +5,7 @@
 
 [Changelog](./CHANGELOG.md) | [Installation](https://sergei-mironov.github.io/litrepl/installation) | [Usage](https://sergei-mironov.github.io/litrepl/usage/general-concepts/) | [Gallery](#-gallery)
 
-[![](./img/coverage.svg)](#-coverage-report)
+[![](./img/coverage.svg)](./docs/coverage.md)
 </div>
 
 <!--
