+++ docs/examples/example.tex
@@ -1,7 +1,7 @@
 \documentclass{article}
 \usepackage{verbatim}
 \usepackage{mdframed}
-\usepackage{minted}
+\usepackage[outputdir=_build]{minted}
 \renewcommand{\MintedPygmentize}{pygmentize}
 \usepackage[utf8]{inputenc}
 
