+++ docs/static/hello.md
@@ -1,5 +1,0 @@
-``` python
-print('Hello Markdown!')
-```
-``` result
-```
