+++ doc/description.md
@@ -1,5 +1,0 @@
-**Litrepl** is a command-line processor for *Markdown* or *LaTeX* documents with
-**literate programming** code sections. Instructed by its arguments, it
-evaluates and updates sections via background interpreters. Interpreters can
-stay active for a **read-eval-paste-loop** style. The repository includes a Vim
-plugin to demonstrate editor integration.
