+++ docs/index.md
@@ -1,7 +1,6 @@
 # Litrepl
 
-**Litrepl** is a command-line processor for *Markdown* or *LaTeX* documents with
-**literate programming** code sections. Instructed by its arguments, it
-evaluates and updates sections via background interpreters. Interpreters can
-stay active for a **read-eval-paste-loop** style. The repository includes a Vim
-plugin to demonstrate editor integration.
+**Litrepl** is a {!static/description.md!}
+
+The [repository](https://github.com/sergei-mironov/litrepl) includes a Vim plugin to demonstrate
+editor integration.
