+++ doc/screencasts/ai.md
@@ -1,91 +1,0 @@
-LitREPL AI Capabilities
-=======================
-
-## Styling
-<!--{{{-->
-
-You can request the AI to rephrase a selected portion
-of the text using the :LAIStyle feature.
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!--}}}-->
-## Requests for code modification
-<!--{{{-->
-
-By this we demonstrate the `:LAICode` command for
-code modification. Consider a program:
-
-```python
-def hello():
-  print(f'Hello, World!')
-```
-
-
-
-
-
-
-
-
-
-<!--}}}-->
-## Correcting the last request
-<!--{{{-->
-
-Litrepl remembers the last AI-modified text region,
-allowing us to apply corrections using the `:LAICont`
-command
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!--}}}-->
-## Direct access to the terminal
-<!--{{{-->
-
-Finally, we can continue the discussion in terminal
-using the `:LAITell` command.
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-<!--}}}-->
-
