+++ data/test-1.md
@@ -1,23 +1,0 @@
-```python
-def hello(name):
-  print(f"Hello, {name}!")
-
-hello("LitREPL")
-```
-
-```
-Hello, LitREPL!
-```
-
-```python
-hello("World")
-```
-
-```
-Hello, World!
-```
-
-* List
-* <!--litrepl-->
-  BLABLA
-  <!--litrepl-->
