+++ data/test-2.md
@@ -1,20 +1,0 @@
-asdasd
-
-v=3
-3+3
-The inline `33=4` snippet
-
-```python
-a=5
-print('Snippet 1', a+3)
-```
-
-```
-zzzzzzzz
-```
-
-
-```python
-b=33
-print('Snippet 2', a+b)
-```
