+++ obs/Litrepl/basic.ini
@@ -1,50 +1,0 @@
-[General]
-Name=Litrepl
-
-[Video]
-BaseCX=442
-BaseCY=361
-OutputCX=442
-OutputCY=361
-FPSType=1
-FPSCommon=10
-ScaleType=bicubic
-ColorRange=Full
-ColorFormat=RGB
-ColorSpace=sRGB
-FPSInt=2
-
-[Panels]
-CookieId=9C6758177BC4D05D
-
-[SimpleOutput]
-RecEncoder=x264
-RecQuality=Small
-FilePath=/home/grwlf/dwnl
-
-[Output]
-Mode=Advanced
-
-[AdvOut]
-TrackIndex=1
-RecType=FFmpeg
-RecSplitFileType=Time
-RecTracks=1
-FLVTrack=1
-FFOutputToFile=true
-FFFormat=gif
-FFFormatMimeType=image/gif
-FFExtension=gif
-FFVBitrate=2400
-FFVEncoderId=97
-FFVEncoder=gif
-FFAEncoderId=0
-FFAEncoder=
-FFAudioMixes=1
-VodTrackIndex=2
-FFVGOPSize=250
-RescaleRes=1920x1080
-RecRescaleRes=1920x1080
-FFRescaleRes=1920x1080
-FFFilePath=/home/grwlf/proj/litrepl.vim/obs
-FFIgnoreCompat=false
