### 2025-05-12 f0a420b
Introduce auxdir parceable tags
