### 2025-05-13 9ae84f2
Raise an error on double-start
