### 2025-05-28 2aabaec
Remove aicli from the flake. Questionable, but will overcome mutual dependency
