### 2025-06-10 #8384 @editorialbot
[REVIEW]: Litrepl: Literate Paper Processor Promoting Transparency More Than Reproducibility
**Submitting author:** @sergei-mironov (<a href="http://orcid.org/0009-0005-1604-722X">Sergei Mironov</a>)
**Repository:** https://github.com/sergei-mironov/litrepl
**Branch with paper.md** (empty if default branch): paper
**Version:** v3.12.0
