### 2025-06-15 9b125bd
Update auxdir syntax; Version 3.13.1
