### 2025-06-15 ac07089
Update auxdir syntax; Version 3.13.1
