### 2025-07-01 sergei-mironov/litrepl#4 @itepifanio
[JOSS Review] Suggested revisions on installation
For the JOSS installation check of [@joss-review](https://github.com/openjournals/joss-reviews/issues/8384)

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Installation works [as documented](https://sergei-mironov.github.io/litrepl/installation/), but there’s some enhancements that can be made:

1. If the author’s prefered installation method is using Nix, why it’s the docs third installation method? I would expected to see the Nix as the first option if that's the case.
2. The README clearly says the socat it's an optional requirement needed for Vim's LTerm to work, but the installation page lacks this explanation.
3. I would suggest adding the extracting the tar file step to the second bullet point at `Installing release versions from Pypi and [Vim.org](http://vim.org/)` subsection (see image below). I had never installed a vim plugin before and my first attempt was to wget the file to the `~/.vim/plugin` without untar it.

![image.png](attachment:82f8acf4-61d6-4150-ba35-aac63edb0563:image.png)

4. There’s a typo in the installation notes subsection `pacakge` instead of `package`
5. The `aicli` interpreter mention at the installation page it's confusing, I had to go back to check if I was missing a installation step. Can you add more context about the `aicli` usage? Since this step is a introduction to your tool, I think this context will add great value to the docs
