### 2025-07-01 sergei-mironov/litrepl#5 @itepifanio
[JOSS Review] Suggested revisions on automated tests
For the JOSS automated tests check of https://github.com/openjournals/joss-reviews/issues/8384

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

There are no [automated tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) in the CI. It doesn’t seem that the build docs is executed via Github Actions. The docs it’s been build using Litrepl, which would count as automated tests once the docs is build, but it doesn’t seem that the Github Pages are build via CI either.

I would also encourage adding a Github Actions that runs tests over all Litrepl supported python versions. Currently it’s not clear which interpreter versions you support, your `[setup.py](http://setup.py/)` says `python >=3.6` but I can’t run the minimal example (`cat file.md | litrepl eval-sections > result.md`) using python 3.6 I notice the following exception:

```python
[GCC 12.2.0] on linux
>>> from dataclasses import dataclass
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ModuleNotFoundError: No module named 'dataclasses'
```

About the coverage, I consider it as a good metric to show where to focus testing efforts. Given your [latest report](https://sergei-mironov.github.io/litrepl/coverage/) I would suggest taking a look at the `eval` module and check if it's worth it to increase your coverage.
