### 2025-07-01 sergei-mironov/litrepl#6 @itepifanio
[JOSS Review] Suggested revisions on code quality
For suggested revisions for https://github.com/openjournals/joss-reviews/issues/8384

I notice that you’re using python `assert` keyword to validate business logic such as `assert ss is not None, f"Failed to attach to {st2name(st)} interpreter”` , which it's not recommend. If python is executed in optimize mode those assert validation will be ignored: https://docs.python.org/3/reference/simple_stmts.html#the-assert-statement

I would also suggest using [pep8 code style](https://peps.python.org/pep-0008/)(https://peps.python.org/pep-0008/) at your code, it’s an approach that is followed by the Python community and would make the codebase more welcoming to contributors.

I also notice that sometimes you use snake_case and other times camelCase, so I search for some guidelines about the code style and didn't find any. Adding a contributing.md or adding [more context to the development section](https://sergei-mironov.github.io/litrepl/development/) would help new contributors.
