### 2025-07-01 sergei-mironov/litrepl#7 @itepifanio
[JOSS Review] Suggested revisions on documentation
For the JOSS documentation check of https://github.com/openjournals/joss-reviews/issues/8384

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Your core features are well documented, the API it’s clear, but I would suggest showing more values at the running examples. The `Vimtex integration (Show Video)` really displays why I would use Litrepl, and I think that the introduction docs lacks some real world examples to show to users the Litrepl usage value.
