### 2025-07-01 sergei-mironov/litrepl#8 @itepifanio
[JOSS Review] Suggested revisions on the paper
For suggested revisions for  https://github.com/openjournals/joss-reviews/issues/8384

There are some structure enhancements that can be made on the paper.

- First the `Figure 1` is displayed but never explained in the text [1]
- The table that compares Litrepl with other tools is not enumerated and referenced directly, just mentioned `The difference between Litrepl and other solutions, ..., is highlighted in the table below` [1]
- I think that the comparison of Litrepl with other tools can provide more value to your text, right now there's only a table comparing, you're not directly arguing why your tool is better or concluding anything from it.
- Your summary can be improved with a clear statement of need of the tool. Right now it says it's inspired by Project Jupyter and then it says it's focused on reproductibility by eliminating mutable states, if you directly tell to the reader which gap Litrepl addresses the text will be more compeling
- I would rethink using the phrase `In the light of recent developments in software deployment` which is too broad and doesn't seem to add much value to your text

[[1] JOSS Internal references](https://joss.readthedocs.io/en/latest/paper.html#internal-references)
