### 2025-08-05 #8384 @matthewfeickert
@sergei-mironov Can you please follow up with @itepifanio on the issues that have been raised in the review so far?

* https://github.com/sergei-mironov/litrepl/issues/4
* https://github.com/sergei-mironov/litrepl/issues/5
* https://github.com/sergei-mironov/litrepl/issues/6
* https://github.com/sergei-mironov/litrepl/issues/7
* https://github.com/sergei-mironov/litrepl/issues/8
