### 2025-08-09 sergei-mironov/litrepl#9 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on installation
Regarding the JOSS review criteria concerning installation outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  Installation: Does installation proceed as outlined in the documentation?

I started by testing out the suggested NixOS installation modes. It was encouraging to discover detailed NixOS instructions, and it is clear that you are confident with Nix packaging approaches — this was excellent to see! That said, I did find it somewhat difficult to actually build the package using the given documentation:

## Building `vim-demo`
Identifying the proper tag to build the `vim-demo` from the repository was difficult; the `vim-demo` derivation pulls the package from PyPI rather than the local filesystem, which introduces a cyclic version number dependency. This is handled by modifying the hash of the sdist in `default.nix`, but this hasn't been done in lock-step with modifying the version number in `semver.txt`.

I am not an expert in Nix packaging best-practice, but I have two suggestions for improving this:
1. Given that you're pulling the source distribution I'd perhaps recommend using the local build (non PyPI) for your demo.
2. Always following the tagged release commit with a commit that modifies the SHA of the PyPI package derivation _immediately_ as the next commit.

I tested the build for 3.13.0 using this SHA: 2aabaec398308f01f9c3b9e6d76ac363069aec3f. The build succeeded, but running `:%LEvel` in the resulting vim distribution did not succeed: for [the first example](https://sergei-mironov.github.io/litrepl/usage/general-concepts/) in the docs, I encountered the following:
```
REPORT
======

litrepl:     bash: --version: command not found
litrepl.vim: 3.13.0+g2aabaec

COMMAND
-------
%!cd "/tmp";  --python-interpreter="auto" --ai-interpreter="auto" --pending-exitcode=33 --debug=0 --filetype=markdown --map-cursor=6:1:/tmp/litrepl_cursor.txt --result-textwidth=53  --timeout=inf,inf  --propagate-sigint eval-sections 1..$ 2>>/tmp/litrepl.err

STDIN
------
<vim-buffer>

STDERR
------
bash: --python-interpreter=auto: command not found
<end-of-stderr>
```

## Building `litrepl-release`
I switched to the original tag for this review: 3.12.0.

Running `nix build .#litrepl-release` successfully built (into `result`). There are many build targets here: given that you enumerate them in [the development section](https://sergei-mironov.github.io/litrepl/development/), I'd link to that from your installation guide.

## Installation documentation
I'd suggest using block-level code rather than inline code for shell program listings, e.g. using Markdown's code-fence markup. This will improve the readability of the `Installing latest versions from Git using Pip and Vim-Plug` section.
