### 2025-08-09 grwlf/litrepl.vim#10 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on documentation to introduce a statement of need
Regarding the JOSS review criteria concerning a statement of need in the documentation outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

I think this could be clearer. The statement of need in your paper is likely more than sufficient! What is in the documentation right now is quite technical; I'd like to see more dialogue around the kinds of people you think might use this tool, and what problem it would solve for them.
