### 2025-08-09 sergei-mironov/litrepl#11 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on documentation RE "Functionality documentation"
Regarding the JOSS review criteria concerning installation outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

You provide sufficient documentation that I felt comfortable trying to use the tool. I would consider splitting your documentation into different sections, as you have two different tools here:
- A core package
- A vim plugin

I would also consider [the Diátaxis framework](https://diataxis.fr/) for structuring your documentation. It can be applied iteratively (as you keep writing docs), and is brings a helpful structure for new users to navigate your content. For example "Application Scenarios" feels a little bit like "How-Tos". I'd also consider splitting out Development and Related into their own top-level section(s) besides USAGE; the USAGE section feels quite big for the moment.

There are small tweaks that might be worth making. For example, the `CLASS` argument to the `repl` command is documented as defaulting to Python, but omitting it raises an error. To me as a reader, that feels surprising — is there really a default?
