### 2025-08-09 sergei-mironov/litrepl#12 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on example usage
Regarding the JOSS review criteria concerning installation outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

When I first saw your documentation, I was hoping to find a quick-start example that showed a real-world example (more than a few lines), and how you might use `litrepl` with it. Really, I think I'm suggesting a quickstart docs entry that allows people to copy-and-paste, including basic installation instruction links. This tutorial can do a few different things, but should be less than 1–2m to run through (in my view).

The purpose is just to get potential users up-and-running before they dig into the more detailed docs.
