### 2025-08-09 sergei-mironov/litrepl#13 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on automated testing
Regarding the JOSS review criteria concerning automated tests outlined in https://github.com/openjournals/joss-reviews/issues/8384
> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

It's helpful that you have a `runtests.sh` helper. You should update the documentation to locate this file (it's in `sh/`). I ran this inside the Nix devShell, and the tests were executed. I saw a test failure:
```
Running test "test_parse_print" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_parse_print" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_eval_md" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "/nix/store/5jw69mbaj5dg4l2bj58acg3gxywfszpj-bash-5.2p26/bin/sh"
Running test "test_eval_md" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "/nix/store/5jw69mbaj5dg4l2bj58acg3gxywfszpj-bash-5.2p26/bin/sh"
Running test "test_tqdm" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_tqdm" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_eval_tex" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "/nix/store/5jw69mbaj5dg4l2bj58acg3gxywfszpj-bash-5.2p26/bin/sh"
Running test "test_eval_tex" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "/nix/store/5jw69mbaj5dg4l2bj58acg3gxywfszpj-bash-5.2p26/bin/sh"
Running test "test_async" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_async" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_eval_code" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_eval_code" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_eval_with_empty_lines" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_eval_with_empty_lines" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_print_system_order" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_print_system_order" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_exit_errcode" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_exit_errcode" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_exception_errcode" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
Running test "test_exception_errcode" python "/nix/store/lks4syng4rmg0zpqlk6zxbw4h1n5jlx9-python3-3.11.9-env/bin/ipython" ai "-" sh "-"
Running test "test_vim_leval_cursor" python "/nix/store/4rf5qybw37b4lh1g0xczlv14sqdbmnpm-python3-3.11.9/bin/python" ai "-" sh "-"
FAIL(127)
```
