### 2025-08-09 grwlf/litrepl.vim#14 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on documentation
In addition to https://github.com/sergei-mironov/litrepl/issues/11 (c.f. https://github.com/openjournals/joss-reviews/issues/8384), I have some additional suggestions regarding your documentation.

In general, you have clearly made an effort with documenting this tool, which is no mean feat given the number of features you've implemented. I want to take a second to acknowledge the work you've put in ­— nicely done!

I would consider [the Diátaxis framework](https://diataxis.fr/) for structuring your documentation. It can be applied iteratively (as you keep writing docs), and is brings a helpful structure for new users to navigate your content. For example "Application Scenarios" feels a little bit like "How-Tos". I'd also consider splitting out Development and Related into their own top-level section(s) besides USAGE; the USAGE section feels quite big for the moment.
