### 2025-08-09 grwlf/litrepl.vim#15 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on contribution guidelines
Regarding the JOSS review criteria concerning community guidelines for contribution outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I'd recommend adding a section on this, so that the barrier for contribution from new users is small. The convention is to add a `CONTRIBUTING.md` file to your GitHub repository. You can then include that in your documentation build via includes with Sphinx using e.g. the `myst-parser` extension for parsing Markdown files.

This usually would be placed in a new section in your documentation, which I briefly touched upon in https://github.com/sergei-mironov/litrepl/issues/11
