### 2025-08-09 grwlf/litrepl.vim#16 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on paper summary
Regarding the JOSS review criteria concerning the software paper summary outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

I think your summary is effective at describing what your tool does. I think it might benefit from a little bit more clarity on the following section:

> In the light of recent developments in software deployment, however, we have shifted our focus from informal reproducibility to enhancing transparency in communication with programming language interpreters, by either eliminating or clearly exposing mutable states within the communication process. The tool is designed to ensure easy code editor integration, and the project repository offers a reference Vim plugin.

Firstly,
> however, we have shifted our focus from informal reproducibility to ...

Here you're referring to "the past" without citing an other publication. This is confusing to readers who do not have the context of who you are or what you've been working on. Unless you've already published another narrative about the problem of reproducibility, I'd suggest not referring to it here.

To follow, this fragment
> by either eliminating or clearly exposing mutable states within the communication process.

doesn't quite make sense to me. When I think about mutable states with _interpreters_, I think about how REPLs are inherently stateful. But, I don't quite know whether that's what you're talking about here, or how this tool solves that problem. For one thing, your later citation of the NixOS thesis makes me think I'm missing something. Would you be able to clarify this part?

Overall, I think you would also benefit from adding more language around the purpose of this tool, e.g. talking about "computational narratives" — why do people want to execute code in text files, and insert the results?
Specifically, I think to address this extract:
> This functionality is useful for both batch document section evaluation
