### 2025-08-09 sergei-mironov/litrepl#17 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on paper state of the field
Regarding the JOSS review criteria concerning documentation of functionality outlined in https://github.com/openjournals/joss-reviews/issues/8384

>  State of the field: Do the authors describe how this software compares to other commonly-used packages?

Whilst you provide a comprehensive table of tools that are related to your project, I think ti would be helpful to add more detail about how these tools differ, and what their focus is.

For example, Quarto is also a publishing tool (exporting to various formats), etc., whilst Jupyter _Lab_ or _Notebook_ are frontends that interact with kernels but do not natively write back to text-based formats like Markdown.

It's possibly worth noting that there are many different kinds of _personas_ that use these tools, and they might have different focuses (e.g. someone who cares about reproducible narratives vs interactive dashboards powered by Markdown notebooks, etc).

---

**Disclaimer: I am a core contributor to the Jupyter Book project.**

Reading your paper, I think it would be useful to mention the Jupyter Book project. It is also a tool that addresses some of the same questions that you refer to, and would sit alongside other tools that you reference like Quarto.
