### 2025-08-09 sergei-mironov/litrepl#18 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on paper references
Regarding the JOSS review criteria concerning paper references outlined in https://github.com/openjournals/joss-reviews/issues/8384
> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

I see a few instances where it would be helpful to combine separate references of the form
```
... APL communities [@Spence1975apl], [@McCarthy1959recfun], [@Iverson1962apl]
```
as
```
... APL communities [@Spence1975apl; @McCarthy1959recfun; @Iverson1962apl]
```
