### 2025-08-09 sergei-mironov/litrepl#19 @agoose77
[JOSS Review, @agoose77]: Suggested revisions on paper statement of need
Regarding the JOSS review criteria concerning the software paper statement of need outlined in https://github.com/openjournals/joss-reviews/issues/8384
>  A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

Overall this feels well-rounded. I think it would benefit from more directly establishing what problem this package is solving, and who the target audience are. Right now, the statement of need feels more like a (very interesting and relevant) historical summary rather than a response to the question of "what problem are we solving, and who are we solving it for?".
