### 2025-08-09 e3768ab
Replace litrepl_extras.vim with script commands
