+++ sh/litrepl-dummy.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+{
+echo $(basename "$0") "$@"
+cat
+echo END
+} | tee _last_dummy_output.txt
