### 2025-08-11 555bc79
Update python: setup prompts using channel
