### 2025-08-11 36fd75b
[HACK] Disable text_extras
