### 2025-08-19 464137f
Re-enable test_vim_extras
