### 2025-08-19 66a039e
IPython: omit log and make initial configuration in-stream
