### 2025-08-19 a5910d3
tests: Remove unused code
