### 2025-08-24 cb2cdf3
Improve error reporting
