+++ python/litrepl/base.py
@@ -27,9 +27,9 @@ from subprocess import check_output, DEVNULL, CalledProcessError
 from contextlib import contextmanager
 
 from .types import (PrepInfo, RunResult, NSec, FileName, SecRec, FileNames,
-                    CursorPos, ReadResult, SType, LitreplArgs,
-                    EvalState, ECode, ECODE_OK, ECODE_RUNNING, SECVAR_RE,
-                    Interpreter, LarkGrammar, Symbols, LarkTree, ParseResult)
+                    CursorPos, ReadResult, SType, LitreplArgs, EvalState,
+                    ECode, ECODE_OK, ECODE_RUNNING, SECVAR_RE, Interpreter,
+                    LarkGrammar, Symbols, LarkTree, ParseResult, ErrorMsg)
 from .eval import (process, pstderr, rresultLoad, rresultSave, processAdapt,
                    processCont, interpExitCode, readipid, with_parent_finally,
                    with_fd, read_nonblock, eval_code, eval_code_, interpIsRunning)
@@ -128,13 +128,12 @@ def pipenames(a:LitreplArgs, st:SType)->FileNames:
                    join(auxdir,"in.pipe"),join(auxdir,"out.pipe"),
                    join(auxdir,"pid.txt"),join(auxdir,"ecode.txt"),join(auxdir,"emsg.txt"))
 
-def attach(fns:FileNames, st:Optional[SType]=None)->Optional[Interpreter]:
+def attach(fns:FileNames, st:Optional[SType]=None)->Union[Interpreter,ErrorMsg]:
   """ Attach to the interpreter associated with the given pipe filenames. """
+  pid=readipid(fns)
+  if pid is None:
+    return f"Could not determine pid of an interpreter"
   try:
-    pid=readipid(fns)
-    if pid is None:
-      pdebug(f"Could not determine pid of an interpreter")
-      return None
     p=Process(pid)
     cmd=p.cmdline()
     cls=None
@@ -147,12 +146,11 @@ def attach(fns:FileNames, st:Optional[SType]=None)->Optional[Interpreter]:
     elif (st is None or st==SType.SShell) and any('sh' in w for w in cmd):
       cls=ShellInterpreter
     else:
-      assert False, f"Unknown or undefined interpreter {cmd} (among {st})"
+      return f"Unknown or undefined interpreter pid {pid} cmd '{cmd}' (among {st})"
     pdebug(f"Interpreter pid {pid} cmd '{cmd}' was resolved into '{cls}'")
     return cls(fns)
   except NoSuchProcess as err:
-    pdebug(f"Could not determine the interpreter classs ({err})")
-    return None
+    return f"Could resolve pid {pid} into interpreter class: ({err})"
 
 def open_child_pipes(inp,outp):
   return os.open(inp,os.O_RDWR|os.O_SYNC),os.open(outp,os.O_RDWR|os.O_SYNC);
@@ -161,6 +159,9 @@ def open_parent_pipes(inp,outp):
 
 
 def write_child_pid(pidf,pid):
+  """ Writes first children of a fork-child process as the main interpreter
+  pid. This works as long as `Interpreter.run_child` calls real interpreter
+  using system() call."""
   with open(pidf,'w') as f:
     for attempt in range(20):
       try:
@@ -466,20 +467,37 @@ def parse_(a:LitreplArgs)->ParseResult:
     raise ValueError(f"Unsupported filetype \"{a.filetype}\"")
   return res
 
+def failmsg(fns:FileNames,ss:Union[Interpreter,str],ec:ECode)->str:
[…]
