### 2025-08-25 3357333
Improve error reporting
