### 2025-09-04 5862551
Fix rebase problem in tex com section test
