### 2025-09-04 c274be4
Implement print-regexp command
