### 2025-09-10 37a6ac9
Fix Makefile man command merge issue
