### 2025-09-10 sergei-mironov/litrepl#4 @sergei-mironov
Hi @itepifanio , Thank you for raising the issue, let me address it. For your first question,

> If the author’s prefered installation method is using Nix, why it’s the docs third installation method?

I put it down the list because Nix is not very popular among Linux users and real Nix users would probably know what to do anyways. So, I think it would be better to keep traditional methods above other, but let me know if you are not convinced.

I am about to address other points very soon.
