### 2025-09-10 1cd8b9a
Update docs: add socat notes, add vim archive unpacking notes, fix a typo, add aicli note #4
