### 2025-09-10 9ba7ca5
Update docs: add socat notes, add vim archive unpacking notes, fix a typo, add aicli note #4
