+++ .github/workflows/run-tests.yaml
@@ -1,0 +1,25 @@
+name: Run Litrepl tests
+
+on: [push]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.10", "3.12"]
+
+    steps:
+      - uses: actions/checkout@v5
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Run the tests
+        run: |
+          python -c "import sys; print(sys.version)"
+          python setup.py install
+
+
