### 2025-09-11 01ef256
Update install_deps_ubuntu.sh
