+++ sh/install_deps_ubuntu.sh
@@ -1,0 +1,4 @@
+#!/bin/sh
+set -e -x
+
+apt-get install python3-setuptools socat
