+++ sh/install_deps_ubuntu.sh
@@ -1,4 +1,5 @@
 #!/bin/sh
 set -e -x
 
-apt-get install python3-setuptools socat
+apt-get update >/dev/null
+apt-get install -y python3-setuptools socat
