+++ sh/install_deps_ubuntu.sh
@@ -2,4 +2,4 @@
 set -e -x
 
 apt-get update >/dev/null
-apt-get install -y python3-setuptools socat
+apt-get install -y python3-setuptools socat vim
