### 2025-09-11 996d479
Update install_deps_ubuntu.sh
