### 2025-09-11 275be5a
Update install_deps_ubuntu.sh
