+++ sh/install_deps_ubuntu.sh
@@ -3,4 +3,3 @@ set -e -x
 
 apt-get update >/dev/null
 apt-get install -y python3-setuptools socat vim
-pip install coverage coverage_badge
