+++ .github/workflows/run-tests.yaml
@@ -20,7 +20,7 @@ jobs:
       - name: Run the tests
         run: |
           python -c "import sys; print(sys.version)"
-          sudo ./sh/install_deps_ubuntu.sh
+          ./sh/install_deps_ubuntu.sh
           python setup.py install
           LITREPL_COVERAGE=- make test
 
