### 2025-09-11 ecf8e1a
Update install_deps_ubuntu.sh
