+++ sh/install_deps_ubuntu.sh
@@ -3,4 +3,4 @@ set -e -x
 
 sudo apt-get update >/dev/null
 sudo apt-get install -y socat vim
-pip install setuptools tqdm coverage coverage_badge
+pip install setuptools tqdm coverage coverage_badge lark psutil
