+++ .github/workflows/run-tests.yaml
@@ -22,6 +22,6 @@ jobs:
           python -c "import sys; print(sys.version)"
           ./sh/install_deps_ubuntu.sh
           python setup.py install
-          ./sh/runtests.sh -c - -p $(which python3) -i $(which python3)
+          ./sh/runtests.sh -c - -p $(which python) -i $(which python)
 
 
