+++ .github/workflows/run-tests.yaml
@@ -22,6 +22,7 @@ jobs:
           python -c "import sys; print(sys.version)"
           ./sh/install_deps_ubuntu.sh
           python setup.py install
-          ./sh/runtests.sh -c - -p $(which python) -i $(which python)
+          export P=`which python`
+          ./sh/runtests.sh -c - -p $P -i $P
 
 
