### 2025-09-12 33c9907
Update runtests.sh
