### 2025-09-13 a9bbb97
Fix test selection logic; Add --dru-run test argument; Add test-small Makefile rule
