### 2025-09-13 8c1b76d
Fix test selection logic; Add --dru-run test argument; Add test-small Makefile rule
