### 2025-09-13 c964df5
Update CI rule, call test-small
