### 2025-09-13 cd5b888
Update CI rule, run python 3.6
