+++ .github/workflows/run-tests.yaml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.10", "3.11"]
+        python-version: ["3.6", "3.10", "3.11"]
 
     steps:
       - uses: actions/checkout@v5
