### 2025-09-13 8d79b43
Update CI rule, run python 3.6
