### 2025-09-13 60d965a
Update CI rule, run python 3.6
