### 2025-09-14 6655543
Rename workflow, remove python 3.6
