+++ .github/workflows/run-tests.yaml
@@ -1,46 +1,0 @@
-name: Run Litrepl tests
-
-on: [push]
-
-jobs:
-  test1:
-    runs-on: ubuntu-20.04
-    strategy:
-      matrix:
-        python-version: ["3.6.15"]
-
-    steps:
-      - uses: actions/checkout@v5
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Run the tests
-        run: |
-          python -c "import sys; print(sys.version)"
-          ./sh/install_deps_ubuntu.sh
-          python setup.py install
-          make test-small
-
-  test2:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.10", "3.11"]
-
-    steps:
-      - uses: actions/checkout@v5
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Run the tests
-        run: |
-          python -c "import sys; print(sys.version)"
-          ./sh/install_deps_ubuntu.sh
-          python setup.py install
-          make test-small
-
-
