+++ .github/workflows/testing.yaml
@@ -1,0 +1,30 @@
+name: Run Litrepl tests
+
+on: [push]
+
+jobs:
+
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
+
+    steps:
+      - uses: actions/checkout@v5
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Install test pre-requisites
+        run: |
+          python -c "import sys; print(sys.version)"
+          ./sh/install_deps_ubuntu.sh
+          python setup.py install
+
+      - name: Run the tests
+        run: |
+          make test-small
+
+
