### 2025-09-20 5295b83
Add ipython to the make test-small
