### 2025-09-20 00028fd
Replace assert with the assert_() function #6
