### 2025-09-20 03095a0
Replace assert with the assert_() function #6
