+++ python/litrepl/eval.py
@@ -245,7 +245,7 @@ def process(a:LitreplArgs,fns:FileNames, ss:Interpreter, lines:str)->Tuple[str,R
   """ Evaluate `lines` synchronously. """
   pdebug("process started")
   p1,p2=ss.patterns()
-  runr=processAsync(fns,ss,lines)
+  runr=process_async(fns,ss,lines)
   res=''
   with with_sigint(a.propagate_sigint,fns):
     with with_locked_fd(runr.fname,OPEN_RDONLY,LOCK_EX) as fdr:
@@ -256,7 +256,7 @@ def process(a:LitreplArgs,fns:FileNames, ss:Interpreter, lines:str)->Tuple[str,R
       pdebug("process readout complete")
   return res,runr
 
-def interpIsRunning(fns:FileNames)->bool:
+def interp_is_running(fns:FileNames)->bool:
   """ Is the background interpreter session running?"""
   try:
     ipid=readipid(fns)
@@ -268,15 +268,15 @@ def interpIsRunning(fns:FileNames)->bool:
       return False
   return True
 
-def interpExitCode(fns:FileNames,
-                   poll_sec=0.5,
-                   poll_attempts=4,
-                   undefined=ECODE_UNDEFINED)->ECode:
+def interp_exitcode(fns:FileNames,
+                    poll_sec=0.5,
+                    poll_attempts=4,
+                    undefined=ECODE_UNDEFINED)->ECode:
   """ Determines retcode of the interpreter. Polls it a little if it looks
   crashed."""
   ecode:int|None=None
   while ecode is None:
-    if interpIsRunning(fns):
+    if interp_is_running(fns):
       return ECODE_RUNNING
     else:
       try:
@@ -343,14 +343,14 @@ OPEN_RDONLY=os.O_RDONLY|os.O_SYNC
 LOCK_NONBLOCKING=LOCK_EX|LOCK_NB
 LOCK_BLOCKING=LOCK_EX
 
-def processAsync(fns:FileNames, ss:Interpreter, code:str)->RunResult:
+def process_async(fns:FileNames, ss:Interpreter, code:str)->RunResult:
   """ Send `code` to the interpreter and fork the response reader. The output
   file is locked and its name is saved into the resulting `RunResult` object.
   """
   wd,inp,outp,_,_,_=astuple(fns)
   codehash=hashdigest(code)
   fname=join(wd,f"partial_{codehash}.txt")
-  pdebug(f"processAsync starting via {fname}")
+  pdebug(f"process_async starting via {fname}")
   with with_locked_fd(fname,CREATE_WRONLY_EMPTY,LOCK_NONBLOCKING) as fo:
     if fo is not None:
       sys.stdout.flush(); sys.stderr.flush() # FIXME: crude
@@ -361,18 +361,18 @@ def processAsync(fns:FileNames, ss:Interpreter, code:str)->RunResult:
         def _handler(signum,frame):
           pass
         signal(SIGINT,_handler)
-        pdebug(f"processAsync reader opening pipes")
+        pdebug(f"process_async reader opening pipes")
         with with_locked_fd(inp, os.O_WRONLY|os.O_SYNC,
                             fcntl.LOCK_EX|fcntl.LOCK_NB,open_timeout_sec=0.5) as fdw:
           with with_locked_fd(outp, os.O_RDONLY|os.O_SYNC,
                               fcntl.LOCK_EX|fcntl.LOCK_NB,open_timeout_sec=0.5) as fdr:
             if fdw and fdr:
-              pdebug("processAsync reader interact start")
+              pdebug("process_async reader interact start")
               try:
                 interact(fdr,fdw,code,fo,ss)
-                pdebug("processAsync reader interact finish")
+                pdebug("process_async reader interact finish")
               except BrokenPipeError:
[…]
