### 2025-09-20 c13febe
Update development.md, add the Contributing section #6
