### 2025-09-20 201b688
Update development.md, add the Contributing section #6
