### 2025-09-20 c768db2
Fix make docs warning
