### 2025-09-20 ab4841d
Fix make docs warning
