### 2025-09-27 sergei-mironov/litrepl#7 @sergei-mironov
Good point, I have added the video to the Basic concepts page.
