### 2025-10-11 621aaf0
Update paper.md, reword the introduction part #8
