### 2025-10-11 9c693f9
Update paper.md, reword the introduction part #8
