### 2025-10-11 a02c6a2
Update paper.md, listed POSIX interfaces #8
