### 2025-10-11 17db6fd
Update paper.md, listed POSIX interfaces #8
