### 2025-10-11 9e00013
Update docs, describe remote SSH use-case
