+++ docs/development.md
@@ -156,6 +156,7 @@ docs:         Build the MkDocs documentation
 examples:     Build examples
 help:         Print help
 man:          Build a manpage
+paper-md:     Check and compile the paper PDF out of its Markdown source using JOSS tools
 paper-quick:  Compile the paper PDF out of its LaTeX source without re-evaluation
 paper:        Check and compile the paper PDF out of its LaTeX source
 readme:       Update code sections in the README.md
