### 2025-10-11 f94dc6d
Update paper.md, describe the Figure #8
