+++ paper/md/paper.md
@@ -33,7 +33,7 @@ plugin.
 
 # Statement of need
 
-![Litrepl resource allocation diagram. Hash **A** is computed based on the Litrepl working directory and the interpreter class. Hash **B** is computed based on the contents of the code section.](./pic.png)
+![\label{figure} Litrepl resource allocation diagram. Hash **A** is computed based on the Litrepl working directory and the interpreter class. Hash **B** is computed based on the contents of the code section.](./pic.png)
 
 
 Literate Programming, formulated by Donald Knuth, shifts the focus from merely
@@ -167,16 +167,13 @@ I use LINUX btw!
 <!--noresult-->
 
 The side effect of this execution is the starting of a session with the Python
-interpreter, which now runs in the background. We communicate with it by editing
-the document and re-running the above command.
-
-Evaluation results are written back into the result sections, and the entire
-document is printed. At this stage, certain conditions can be optionally
-checked. First, adding `--pending-exitcode=INT` instructs Litrepl to report an
-error if a section takes longer than the timeout to evaluate. Second, setting
-`--exception-exitcode=INT` directs Litrepl to detect Python exceptions. Lastly,
-`--irreproducible-exitcode=INT` triggers an error if the evaluation result
-doesn't match the text initially present in the result section.
+interpreter, which now runs in the background. The resources that builds up this
+functionality are illustrated on the Figure \ref{figure}. The session is
+represented by a couple of pipes, the interpreter process identifier, and by a
+sink file for asyncronous response. Litrepl connects to a session by sending the
+contents of a code section to the input pipe and forking out a reader process
+receiving the interpreter response from the output pipe. POSIX file lock
+mechanism helps Litrepl to manage the concurrent access to the pipes.
 
 
 ## Interfacing Interpreters
