### 2025-10-11 74c4d32
Update paper.md, describe the Figure #8
