### 2025-10-12 006141a
Update paper.md, describe the Figure #8
