### 2025-10-12 ec6a8de
Update paper.md, describe the Figure #8
