### 2025-10-12 d6ef003
Update paper.md, Reference the table #8
