### 2025-10-12 98a95fe
Update paper.md, Reference the table #8
