### 2025-10-20 224f66f
Update paper.md, Add direct arguments #8
