### 2025-10-23 a39883d
Update aicli interpreter: require explicit /ask endings
