### 2025-10-31 #8384 @matthewfeickert
@sergei-mironov I see from your commit history (e.g. https://github.com/sergei-mironov/litrepl/commit/a39883d4f3ac058ef27d3455c8f34c2d0774266b) that there is activity on `litrepl`, which is good. Though please take time to communicate with the reviewers on their open Issues (https://github.com/sergei-mironov/litrepl/issues) to let them know your plans for addressing them. I see that you've done some of that on https://github.com/sergei-mironov/litrepl/issues/8, which you resolved, but please communicate in advance with the reviewers to make sure that you're aligned.
