### 2025-11-14 #8384 @matthewfeickert
@sergei-mironov It has been two weeks now and there has been no further activity on the repository since [2025-10-23](https://github.com/sergei-mironov/litrepl/commit/a39883d4f3ac058ef27d3455c8f34c2d0774266b) and there has been no response to any of the remaining [open issues](https://github.com/sergei-mironov/litrepl/issues) that were opened on 2025-08-09. Do you have time to commit to the review currently? Or should we pause the review?
