### 2025-11-19 f3e687b
Print error when called without arguments from a tty
