### 2025-11-19 82ab9c1
Add development vimrc script
