### 2025-11-20 e964311
Make a safer ipython setup
