+++ python/litrepl/interpreters/ipython.py
@@ -24,12 +24,12 @@ class IPythonInterpreter(Interpreter):
       'import sys\n'
       # See https://github.com/ipython/ipython/issues/14246
       'sys.stdout.reconfigure(line_buffering=True)\n'
-      'from IPython.terminal.prompts import Prompts, Token\n'
-      'class EmptyPrompts(Prompts):\n'
+      'import IPython.terminal.prompts as _ipyp\n'
+      'class EmptyPrompts(_ipyp.Prompts):\n'
       '  def in_prompt_tokens(self):\n'
-      '    return [(Token.Prompt, "")]\n'
+      '    return [(_ipyp.Token.Prompt, "")]\n'
       '  def continuation_prompt_tokens(self, width=None):\n'
-      '    return [(Token.Prompt, "") ]\n'
+      '    return [(_ipyp.Token.Prompt, "") ]\n'
       '  def rewrite_prompt_tokens(self):\n'
       '    return []\n'
       '  def out_prompt_tokens(self):\n'
