### 2025-11-20 d174417
Update vim_litrepl_extras.vim: ignore repeating action matches
