### 2025-11-21 ea90062
Recognize IPython cmdlines as ipython
