### 2025-11-24 4557e21
Update vim_litrepl_extras.vim: dont pass --reindent, pass eval-code command explicitly
