### 2025-11-25 dd0bd1c
Add --keep-readout debugging option
