### 2025-11-25 94edb65
Fix leaking stdout in on-demand start mode
