### 2025-11-28 99a4afd
Vim extras: pass --location args with vim registers
