### 2025-12-03 0ea3bc0
Vim extras: pass command as --command
