### 2025-12-03 fa90b0e
Document litrepl_extras.vim
