### 2026-01-05 #8384 @matthewfeickert
@sergei-mironov can you please respond to the track EiC's questions on timelines? https://github.com/openjournals/joss-reviews/issues/8384#issuecomment-3637031532
