### 2026-01-07 #8384 @sergei-mironov
@danielskatz @matthewfeickert Let me review the issues. I will split them in two categories:
(1) Suggestions for the paper text.
- [x] https://github.com/sergei-mironov/litrepl/issues/10
- [x] https://github.com/sergei-mironov/litrepl/issues/19
- [x] https://github.com/sergei-mironov/litrepl/issues/18
- [x] https://github.com/sergei-mironov/litrepl/issues/17
- [x] https://github.com/sergei-mironov/litrepl/issues/16

(2) Suggestions for the code repository.
- [x] https://github.com/sergei-mironov/litrepl/issues/11
- [x] https://github.com/sergei-mironov/litrepl/issues/9
- [x] https://github.com/sergei-mironov/litrepl/issues/15
- [x] https://github.com/sergei-mironov/litrepl/issues/14
- [x] https://github.com/sergei-mironov/litrepl/issues/13 (a bug maybe)
- [x] https://github.com/sergei-mironov/litrepl/issues/12

The suggestions for the paper text will take me 3 weeks I think, Angus made some really strong points in the last two. The second group should be easier, 1-2 weeks. So my final estimate is: 1 month and a half.
