### 2026-02-22 #8384 @sergei-mironov
@danielskatz I hope I addressed all the issues, feel free to review the project. In the process I updated the version from `3.12.0` to `3.14.1`. I also have force-rebased `paper` branch to point to the actual code.
