### 2026-02-22 #8384 @matthewfeickert
👋 @itepifanio, @agoose77 we've had a very long break in this review process, so are you both able and willing to continue reviewing this submission? If so, please check that your issues open on https://github.com/sergei-mironov/litrepl/issues/ were adequately addressed (@sergei-mironov  you closed these issues without getting feedback, so please reopen them if requested given that you closing them keeps the author from reopening the issue themselves). If you're no longer able to continue with the review given other time commitments please email me.
