### 2026-02-23 #8384 @sergei-mironov
>  you closed these issues without getting feedback, so please reopen them if requested given that you closing them keeps the author from reopening the issue themselves

@matthewfeickert I though issue authors should be able to reopen the issues. But I will re-open too if needed.
