### 2026-03-14 #8384 @itepifanio
Hi @sergei-mironov

I’m nearly ready to sign off, but I’d like to see a final pass on some minor typos to polish the text.

I don't have permission to reopen [this issue](https://github.com/sergei-mironov/litrepl/issues/8#issuecomment-3539485324) but the last comment has the specific details about the typos. Once those are addressed, I'm happy to approve!
