### 2026-04-28 #8384 @danielskatz
@matthewfeickert - As track chair, I try to check on submissions in the CSISM track semi-regularly. For this one, what's its status? I can't see that anything has happened in the last month.  But it looks like it might be ready to recommend for acceptance?
