### 2026-04-30 #8384 @matthewfeickert
@itepifanio, @agoose77 thank you for your reviews!

@sergei-mironov As Litrepl looks like it has passed all the reviewer comments I'll review the paper for editorial comments and then we can move forward with publication. Apologies to everyone for the very delayed response time.

In reviewing the paper draft I have created https://github.com/sergei-mironov/litrepl/issues/21 covering editorial issues.
